Module OCanren

Logic values

type 'a logic =

| Var of GT.int * 'a logic GT.list

| Value of 'a

A type of a logic value

class virtual ['ia, 'a, 'sa, 'inh, 'extra, 'syn] logic_t : object ... end

val gcata_logic : < c_Var : 'inh -> 'a logic -> GT.int -> 'a logic GT.list -> 'syn ; c_Value : 'inh -> 'a logic -> 'a -> 'syn.. > -> 'inh -> 'a logic -> 'syn

class ['a, 'extra_logic] fmt_logic_t : ( Stdlib.Format.formatter -> 'a -> unit ) -> ( Stdlib.Format.formatter -> 'a logic -> unit ) -> object ... end

val fmt_logic : ( Stdlib.Format.formatter -> 'a -> unit ) -> Stdlib.Format.formatter -> 'a logic -> unit

class ['a, 'syn, 'extra_logic] foldr_logic_t : ( 'syn -> 'a -> 'syn ) -> ( 'syn -> 'a logic -> 'syn ) -> object ... end

val foldr_logic : ( 'syn -> 'a -> 'syn ) -> 'syn -> 'a logic -> 'syn

class ['a, 'syn, 'extra_logic] foldl_logic_t : ( 'syn -> 'a -> 'syn ) -> ( 'syn -> 'a logic -> 'syn ) -> object ... end

val foldl_logic : ( 'syn -> 'a -> 'syn ) -> 'syn -> 'a logic -> 'syn

class ['a, 'extra_logic] compare_logic_t : ( 'a -> 'a -> GT.comparison ) -> ( 'a logic -> 'a logic -> GT.comparison ) -> object ... end

val compare_logic : ( 'a -> 'a -> GT.comparison ) -> 'a logic -> 'a logic -> GT.comparison

class ['a, 'extra_logic] eq_logic_t : ( 'a -> 'a -> bool ) -> ( 'a logic -> 'a logic -> bool ) -> object ... end

val eq_logic : ( 'a -> 'a -> bool ) -> 'a logic -> 'a logic -> bool

class ['a, 'extra_logic] html_logic_t : ( unit -> 'a -> HTML.er ) -> ( unit -> 'a logic -> HTML.er ) -> object ... end

val html_logic : ( 'a -> HTML.er ) -> 'a logic -> HTML.er

class ['a, 'a_2, 'extra_logic, 'syn_logic] gmap_logic_t : ( unit -> 'a -> 'a_2 ) -> ( unit -> 'a logic -> 'a_2 logic ) -> object ... end

val gmap_logic : ( 'a -> 'a_2 ) -> 'a logic -> 'a_2 logic

class ['a, 'extra_logic] show_logic_t : ( unit -> 'a -> string ) -> ( unit -> 'a logic -> string ) -> object ... end

val show_logic : ( 'a -> string ) -> 'a logic -> string

val logic : ( < c_Var : 'inh -> 'a logic -> GT.int -> 'a logic GT.list -> 'syn ; c_Value : 'inh -> 'a logic -> 'a -> 'syn.. > -> 'inh -> 'a logic -> 'syn, < fmt : ( Stdlib.Format.formatter -> 'a -> unit ) -> Stdlib.Format.formatter -> 'a logic -> unit ; foldr : ( 'syn -> 'a -> 'syn ) -> 'syn -> 'a logic -> 'syn ; foldl : ( 'syn -> 'a -> 'syn ) -> 'syn -> 'a logic -> 'syn ; compare : ( 'a -> 'a -> GT.comparison ) -> 'a logic -> 'a logic -> GT.comparison ; eq : ( 'a -> 'a -> bool ) -> 'a logic -> 'a logic -> bool ; html : ( 'a -> HTML.er ) -> 'a logic -> HTML.er ; gmap : ( 'a -> 'a_2 ) -> 'a logic -> 'a_2 logic ; show : ( 'a -> string ) -> 'a logic -> string >, ( ( 'inh2 -> 'a1 logic -> 'syn3 ) -> [ 'a1_i, 'a1, 'a1_s, 'inh2, 'a1 logic, 'syn3 ] logic_t ) -> 'inh2 -> 'a1 logic -> 'syn3 ) GT.t

val to_logic : 'a -> 'a logic

to_logic x makes a logic value from a regular one

val from_logic : 'a logic -> 'a

from_logic x makes a regular value from a logic one. Raises exception Not_a_value if x contains free variables

Injections/projections

type 'a ilogic

The type ('a, 'b) injected describes an injection of a type 'a into 'b

module Reifier : sig ... end

val inj : 'a -> 'a ilogic

inj x injects x into logical x

val inji : 'a -> 'a ilogic

A synonim to inj. To be removed

val (!!) : 'a -> 'a ilogic

A synonym for fun x -> inj @@ lift x (for non-parametric types)

val prj_exn : ( 'a ilogic, 'a ) Reifier.t

prj_exn x returns a regular value from injected representation. Raises exception Not_a_value if x contains free variables

val prj : ( int -> 'a ) -> ( 'a ilogic, 'a ) Reifier.t

val reify : ( 'a ilogic, 'a logic ) Reifier.t

A default shallow reifier. An alias of Reifier.reify

exception Not_a_value

The exception is raised when we try to extract a regular term from the answer with some free variables

class type 'a reified = object ... end

Reification result

val make_rr : Env.t -> 'a ilogic -> 'a reified

module State : sig ... end

type 'a goal'

Goal converts a state into a lazy stream of states

type goal = State.t Stream.t goal'

miniKanren basic combinators

val call_fresh : ( 'a OCanren__Logic.ilogic -> goal ) -> goal

call_fresh f creates a fresh logical variable and passes it to the parameter

val (===) : 'a OCanren__Logic.ilogic -> 'a OCanren__Logic.ilogic -> goal

x === y creates a goal, which performs a unification of x and y

val unify : 'a OCanren__Logic.ilogic -> 'a OCanren__Logic.ilogic -> goal

unify x y is a prefix synonym for x === y

val (=/=) : 'a OCanren__Logic.ilogic -> 'a OCanren__Logic.ilogic -> goal

x =/= y creates a goal, which introduces a disequality constraint for x and y

val diseq : 'a OCanren__Logic.ilogic -> 'a OCanren__Logic.ilogic -> goal

diseq x y is a prefix synonym for x =/= y

val structural : 'a -> ( 'a -> 'b ) Env.Monad.t -> ( 'b -> bool ) -> goal

Call structural var reifier checker adds a structural constraint for future use. Every time substitution is updated it reifies var using reifier and checks that the result satisfies desired predicate checker.

The predicate checker returns false when constraint is violated. Call structural saves constraint and return substitution nonmodified.

See also: debug_var.

val conj : goal -> goal -> goal

conj s1 s2 creates a goal, which is a conjunction of its arguments

val (&&&) : goal -> goal -> goal

&&& is a left-associative infix synonym for conj

val disj : goal -> goal -> goal

disj s1 s2 creates a goal, which is a disjunction of its arguments

val (|||) : goal -> goal -> goal

||| is a left-associative infix synonym for disj

val (?|) : goal list -> goal

?| [s1; s2; ...; sk] calculates s1 ||| (s2 ||| ... ||| sk)...) for a non-empty list of goals (note the right association)

val conde : goal list -> goal

conde is a synonym for ?|

val (?&) : goal list -> goal

?& [s1; s2; ...; sk] calculates s1 &&& (s2 && ... &&& sk)...) for a non-empty list of goals (note the right association)

Some predefined goals

val success : goal

success always succeeds

val failure : goal

failure always fails

module Fresh : sig ... end

Top-level running primitives

val run : ( unit -> ( 'a -> State.t -> 'b ) * ( 'c -> Env.t -> 'd ) * ( 'b -> 'c * State.t Stream.t ) * ( 'e -> 'd -> 'f ) ) -> 'a -> 'e -> 'f Stream.t

run n g h runs a goal g with n logical parameters and passes reified results to the handler h. The number of parameters is encoded using variadic machinery à la Olivier Danvy and represented by a number of predefined numerals and successor function (see below). The reification replaces each variable, passed to g, with the stream of values, associated with that variable as the goal succeeds.

Examples:

run one (fun q -> q === !!5) (fun qs -> ...). Here qs --- a stream of all values, associated with the variable q.

run two (fun q r -> q === !!5 ||| r === !!6) (fun qs rs -> ...). Here qs, rs --- streams of all values, associated with the variable q and r, respectively.

run (succ one) (fun q r -> q === !!5 ||| r === !!6) (fun qs rs -> ...). The same as the above.

val delay : ( unit -> goal ) -> goal

The primitive delay helps to construct recursive goals, which depend on themselves. For example, we can't write let rec fives q = (q === !!5) ||| (fives q) because the generation of this goal leads to infinite recursion. The correct way to implement this is let rec fives q = (q === !!5) ||| delay (fun () -> fives q)

See also syntax extension defer.

val succ : ( unit -> ( 'a -> State.t -> 'b ) * ( 'c -> Env.t -> 'd ) * ( 'e -> 'f * 'g ) * ( 'h -> 'i -> 'j ) ) -> unit -> ( ( 'k OCanren__Logic.ilogic -> 'a ) -> State.t -> 'k OCanren__Logic.ilogic * 'b ) * ( ('m OCanren__Logic.ilogic * 'c) -> Env.t -> 'm OCanren__Logic.reified * 'd ) * ( ('o * 'e) -> ('o * 'f) * 'g ) * ( ( 'p -> 'h ) -> ('p * 'i) -> 'j )

Successor function

module NUMERAL_TYPS : sig ... end

val one : ( _, _, _, _, _ ) NUMERAL_TYPS.one

Predefined numerals (one to five)

val two : ( _, _, _, _, _, _, _, _, _, _ ) NUMERAL_TYPS.two

val three : ( _, _, _, _, _, _, _, _, _, _, _, _, _, _ ) NUMERAL_TYPS.three

val four : ( _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _ ) NUMERAL_TYPS.four

val q : ( _, _, _, _, _ ) NUMERAL_TYPS.one

The same numerals with conventional names

val qr : ( _, _, _, _, _, _, _, _, _, _ ) NUMERAL_TYPS.two

val qrs : ( _, _, _, _, _, _, _, _, _, _, _, _, _, _ ) NUMERAL_TYPS.three

val qrst : ( _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _ ) NUMERAL_TYPS.four

Tabling primitives. Tabling allows to cache answers of the goal between different queries.

Usage: General form : Tabling.tabled/tabledrec n g where n is the number of parameters and g is the goal. Returns modified `tabled` goal.

1) For non-recursive goals:

let g = Tabling.(tabled two) (fun q r -> q === r)

2) For recursive goals: In this case it is necessery to `abstract` from recursive calls. The goal should take additional (first) argument grec and use it instead of recursive calls to itself.

let g = Tabling.(tabledrec one) (fun grec q -> (q === O) ||| (fresh (n) (q === S n) &&& (grec n)))

module Tabling : sig ... end

val debug_var : 'a OCanren__Logic.ilogic -> ( 'a OCanren__Logic.ilogic -> Env.t -> 'b ) -> ( 'b list -> goal ) -> goal

The call debug_var var reifier callback performs reification of variable var in a current state using reifier and passes list of answer to callback (multiple answers can arise in presence of disequality constraints). The callback can investigate reified value and construct required goal to continue search.

See also: structural.

val only_head : goal -> goal

The goal only_head f returns no answers when f returns:

empty stream when f returns empty stream;

hangs when f hangs during search for first answer;

stream with head answer if f returns stream that has at least 1 answer.

module PrunesControl : sig ... end

module Stream : sig ... end

module Runconf : sig ... end

module Timer : sig ... end

module Env : sig ... end

module Std : sig ... end