Module OCanren.Env

A distinct environment is associated with each OCanren.run. Every fresh variable is associated with an emvironment. Using nested run's with reusing of fresh variables from the previous run will lead to runtime error due to usage of alien environment.

type t

val empty : unit -> t

val create : anchor:int -> t

val fresh : scope:OCanren__Term.Var.scope -> t -> 'a

val check : t -> OCanren__Term.Var.t -> bool

val check_exn : t -> OCanren__Term.Var.t -> unit

val is_var : t -> 'a -> bool

val var : t -> 'a -> OCanren__Term.Var.t option

val freevars : t -> 'a -> OCanren__.Term.VarSet.t

val is_open : t -> 'a -> bool

val equal : t -> t -> bool

module Monad : sig ... end

Essentially, a reader monad over Env.t. Useful for reification.

type 'a m = 'a Monad.t