Module OCanren.Reifier

type ('a, 'b) t = ( 'a -> 'b ) Env.Monad.t

Reifier from type `'a` into type `'b` is an `'a -> 'b` function dipped into the Env.t monad, will see how it plays later.

Perhaps, it is possible to not expose the reifier type and make it itself a monad or something else that composes nicely, but we haven't figured out yet.

A few predefined reifiers from which other reifiers will be composed

val reify : ( 'a ilogic, 'a logic ) t

val prj_exn : ( 'a ilogic, 'a ) t

This one projects implicit logic into the underlying type, raising exception Not_a_value if it finds a variable

val prj : ( int -> 'a ) -> ( 'a ilogic, 'a ) t

val compose : ( 'a, 'b ) t -> ( 'b, 'c ) t -> ( 'a, 'c ) t

Composition of two reifiers

Reifier is a profunctor. Below are two combinators to compose reifiers with regular functions

val fmap : ( 'b -> 'c ) -> ( 'a, 'b ) t -> ( 'a, 'c ) t

val fcomap : ( 'a -> 'b ) -> ( 'b, 'c ) t -> ( 'a, 'c ) t

val fix : ( ( 'a, 'b ) t as 'c -> 'c ) -> 'c

Fixpoint for reifier.

val rework : fv:( 'a Env.m -> 'b Env.m ) -> ( 'a logic Env.m -> 'b logic Env.m ) -> 'a logic Env.m -> 'b logic Env.m

And uxiliary function to make reifiers generated by generic programming point-free

val zed : ( ( 'a -> 'b ) -> 'a -> 'b ) -> 'a -> 'b

Fix point combinator for Call-by-value