Module Std.Bool

Relational booleans

GT-related API

type 'a logic'

Type synonym to prevent toplevel logic from being hidden

class virtual ['ia, 'a, 'sa, 'inh, 'extra, 'syn] logic'_t : object ... end

val gcata_logic' : [ _, 'a, 'sa, 'inh, 'a logic', 'syn ] logic'_t -> 'inh -> 'a logic' -> 'syn

class ['a, 'extra_logic'] fmt_logic'_t : ( Stdlib.Format.formatter -> 'a -> unit ) -> ( Stdlib.Format.formatter -> 'a logic' -> unit ) -> object ... end

val fmt_logic' : ( Stdlib.Format.formatter -> 'a -> unit ) -> Stdlib.Format.formatter -> 'a logic' -> unit

class ['a, 'a_2, 'extra_logic', 'syn_logic'] gmap_logic'_t : ( unit -> 'a -> 'a_2 ) -> ( unit -> 'a logic' -> 'a_2 logic' ) -> object ... end

val gmap_logic' : ( 'a -> 'a_2 ) -> 'a logic' -> 'a_2 logic'

class ['a, 'syn, 'extra_logic'] foldl_logic'_t : ( 'syn -> 'a -> 'syn ) -> ( 'syn -> 'a logic' -> 'syn ) -> object ... end

val foldl_logic' : ( 'syn -> 'a -> 'syn ) -> 'syn -> 'a logic' -> 'syn

class ['a, 'syn, 'extra_logic'] foldr_logic'_t : ( 'syn -> 'a -> 'syn ) -> ( 'syn -> 'a logic' -> 'syn ) -> object ... end

val foldr_logic' : ( 'syn -> 'a -> 'syn ) -> 'syn -> 'a logic' -> 'syn

class ['a, 'extra_logic'] compare_logic'_t : ( 'a -> 'a -> GT.comparison ) -> ( 'a logic' -> 'a logic' -> GT.comparison ) -> object ... end

val compare_logic' : ( 'a -> 'a -> GT.comparison ) -> 'a logic' -> 'a logic' -> GT.comparison

class ['a, 'extra_logic'] eq_logic'_t : ( 'a -> 'a -> bool ) -> ( 'a logic' -> 'a logic' -> bool ) -> object ... end

val eq_logic' : ( 'a -> 'a -> bool ) -> 'a logic' -> 'a logic' -> bool

class ['a, 'extra_logic'] html_logic'_t : ( unit -> 'a -> HTML.er ) -> ( unit -> 'a logic' -> HTML.er ) -> object ... end

val html_logic' : ( 'a -> HTML.er ) -> 'a logic' -> HTML.er

class ['a, 'extra_logic'] show_logic'_t : ( unit -> 'a -> string ) -> ( unit -> 'a logic' -> string ) -> object ... end

val show_logic' : ( 'a -> string ) -> 'a logic' -> string

val logic' : ( [ _, 'a, 'sa, 'inh, 'a logic', 'syn ] logic'_t -> 'inh -> 'a logic' -> 'syn, < fmt : ( Stdlib.Format.formatter -> 'a -> unit ) -> Stdlib.Format.formatter -> 'a logic' -> unit ; gmap : ( 'a -> 'a_2 ) -> 'a logic' -> 'a_2 logic' ; foldl : ( 'syn -> 'a -> 'syn ) -> 'syn -> 'a logic' -> 'syn ; foldr : ( 'syn -> 'a -> 'syn ) -> 'syn -> 'a logic' -> 'syn ; compare : ( 'a -> 'a -> GT.comparison ) -> 'a logic' -> 'a logic' -> GT.comparison ; eq : ( 'a -> 'a -> bool ) -> 'a logic' -> 'a logic' -> bool ; html : ( 'a -> HTML.er ) -> 'a logic' -> HTML.er ; show : ( 'a -> string ) -> 'a logic' -> string >, ( ( 'inh2 -> 'a1 logic' -> 'syn3 ) -> [ 'a1_i, 'a1, 'a1_s, 'inh2, 'a1 logic', 'syn3 ] logic'_t ) -> 'inh2 -> 'a1 logic' -> 'syn3 ) GT.t

type t = GT.bool

Synonym for boolean type

class virtual ['inh, 'extra, 'syn] t_t : object ... end

val gcata_t : [ 'inh, t, 'syn ] t_t -> 'inh -> t -> 'syn

class 'extra_t fmt_t_t : ( Stdlib.Format.formatter -> t -> unit ) -> object ... end

val fmt_t : Stdlib.Format.formatter -> t -> unit

class ['extra_t, 'syn_t] gmap_t_t : ( unit -> t -> t ) -> object ... end

val gmap_t : t -> t

class ['syn, 'extra_t] foldl_t_t : ( 'syn -> t -> 'syn ) -> object ... end

val foldl_t : 'syn -> t -> 'syn

class ['syn, 'extra_t] foldr_t_t : ( 'syn -> t -> 'syn ) -> object ... end

val foldr_t : 'syn -> t -> 'syn

class 'extra_t compare_t_t : ( t -> t -> GT.comparison ) -> object ... end

val compare_t : t -> t -> GT.comparison

class 'extra_t eq_t_t : ( t -> t -> bool ) -> object ... end

val eq_t : t -> t -> bool

class 'extra_t html_t_t : ( unit -> t -> HTML.er ) -> object ... end

val html_t : t -> HTML.er

class 'extra_t show_t_t : ( unit -> t -> string ) -> object ... end

val show_t : t -> string

val t : ( [ 'inh, t, 'syn ] t_t -> 'inh -> t -> 'syn, < fmt : Stdlib.Format.formatter -> t -> unit ; gmap : t -> t ; foldl : 'syn -> t -> 'syn ; foldr : 'syn -> t -> 'syn ; compare : t -> t -> GT.comparison ; eq : t -> t -> bool ; html : t -> HTML.er ; show : t -> string >, ( ( 'inh1 -> t -> 'syn2 ) -> [ 'inh1, t, 'syn2 ] t_t ) -> 'inh1 -> t -> 'syn2 ) GT.t

type ground = GT.bool

Ground boolean (the regular one)

class virtual ['inh, 'extra, 'syn] ground_t : object ... end

val gcata_ground : [ 'inh, ground, 'syn ] ground_t -> 'inh -> ground -> 'syn

class 'extra_ground fmt_ground_t : ( Stdlib.Format.formatter -> ground -> unit ) -> object ... end

val fmt_ground : Stdlib.Format.formatter -> ground -> unit

class ['extra_ground, 'syn_ground] gmap_ground_t : ( unit -> ground -> ground ) -> object ... end

val gmap_ground : ground -> ground

class ['syn, 'extra_ground] foldl_ground_t : ( 'syn -> ground -> 'syn ) -> object ... end

val foldl_ground : 'syn -> ground -> 'syn

class ['syn, 'extra_ground] foldr_ground_t : ( 'syn -> ground -> 'syn ) -> object ... end

val foldr_ground : 'syn -> ground -> 'syn

class 'extra_ground compare_ground_t : ( ground -> ground -> GT.comparison ) -> object ... end

val compare_ground : ground -> ground -> GT.comparison

class 'extra_ground eq_ground_t : ( ground -> ground -> bool ) -> object ... end

val eq_ground : ground -> ground -> bool

class 'extra_ground html_ground_t : ( unit -> ground -> HTML.er ) -> object ... end

val html_ground : ground -> HTML.er

class 'extra_ground show_ground_t : ( unit -> ground -> string ) -> object ... end

val show_ground : ground -> string

val ground : ( [ 'inh, ground, 'syn ] ground_t -> 'inh -> ground -> 'syn, < fmt : Stdlib.Format.formatter -> ground -> unit ; gmap : ground -> ground ; foldl : 'syn -> ground -> 'syn ; foldr : 'syn -> ground -> 'syn ; compare : ground -> ground -> GT.comparison ; eq : ground -> ground -> bool ; html : ground -> HTML.er ; show : ground -> string >, ( ( 'inh1 -> ground -> 'syn2 ) -> [ 'inh1, ground, 'syn2 ] ground_t ) -> 'inh1 -> ground -> 'syn2 ) GT.t

type logic = GT.bool logic'

Logic boolean

class virtual ['inh, 'extra, 'syn] logic_t : object ... end

val gcata_logic : [ 'inh, logic, 'syn ] logic_t -> 'inh -> logic -> 'syn

class 'extra_logic fmt_logic_t : ( Stdlib.Format.formatter -> logic -> unit ) -> object ... end

val fmt_logic : Stdlib.Format.formatter -> logic -> unit

class ['extra_logic, 'syn_logic] gmap_logic_t : ( unit -> logic -> logic ) -> object ... end

val gmap_logic : logic -> logic

class ['syn, 'extra_logic] foldl_logic_t : ( 'syn -> logic -> 'syn ) -> object ... end

val foldl_logic : 'syn -> logic -> 'syn

class ['syn, 'extra_logic] foldr_logic_t : ( 'syn -> logic -> 'syn ) -> object ... end

val foldr_logic : 'syn -> logic -> 'syn

class 'extra_logic compare_logic_t : ( logic -> logic -> GT.comparison ) -> object ... end

val compare_logic : logic -> logic -> GT.comparison

class 'extra_logic eq_logic_t : ( logic -> logic -> bool ) -> object ... end

val eq_logic : logic -> logic -> bool

class 'extra_logic html_logic_t : ( unit -> logic -> HTML.er ) -> object ... end

val html_logic : logic -> HTML.er

class 'extra_logic show_logic_t : ( unit -> logic -> string ) -> object ... end

val show_logic : logic -> string

val logic : ( [ 'inh, logic, 'syn ] logic_t -> 'inh -> logic -> 'syn, < fmt : Stdlib.Format.formatter -> logic -> unit ; gmap : logic -> logic ; foldl : 'syn -> logic -> 'syn ; foldr : 'syn -> logic -> 'syn ; compare : logic -> logic -> GT.comparison ; eq : logic -> logic -> bool ; html : logic -> HTML.er ; show : logic -> string >, ( ( 'inh1 -> logic -> 'syn2 ) -> [ 'inh1, logic, 'syn2 ] logic_t ) -> 'inh1 -> logic -> 'syn2 ) GT.t

Relational API

val inj : ground -> logic

Logic injection (for reification)

type groundi

A synonym for injected boolean; use Logic.inji operator to make a groundi from a regular bool

val reify : ( bool OCanren__Logic.ilogic -> bool OCanren__Logic.logic ) Env.Monad.t

Reifier

val prj_exn : ( bool OCanren__Logic.ilogic -> bool ) Env.Monad.t

Shallow reifier

val falso : groundi

Constants

val truo : groundi

val (|^) : groundi -> groundi -> groundi -> OCanren__Core.State.t Stream.t OCanren__Core.goal'

Sheffer stroke

val noto : groundi -> groundi -> OCanren__Core.State.t Stream.t OCanren__Core.goal'

Negation

val (~~) : groundi -> OCanren__Core.State.t Stream.t OCanren__Core.goal'

Negation as a goal

val oro : groundi -> groundi -> groundi -> OCanren__Core.State.t Stream.t OCanren__Core.goal'

Disjunction

val (||) : groundi -> groundi -> OCanren__Core.State.t Stream.t OCanren__Core.goal'

Disjunction as a goal

val ando : groundi -> groundi -> groundi -> OCanren__Core.State.t Stream.t OCanren__Core.goal'

Conjunction

val (&&) : groundi -> groundi -> OCanren__Core.State.t Stream.t OCanren__Core.goal'

Conjunction as a goal