Module OCanren.Stream

type 'a t

Stream type to represent an infinite sequence of results for a relational query. By convention it is called 'stream' but internally it is more like OCaml's Seq.t, it doesn't do any memoization because in relational programming it is not needed.

Constructors

val nil : 'a t

val single : 'a -> 'a t

val cons : 'a -> 'a t -> 'a t

val from_fun : ( unit -> 'a t ) -> 'a t

Other functions

val suspend : is_ready:( unit -> bool ) -> ( unit -> 'a t ) -> 'a t

val of_list : 'a list -> 'a t

val is_empty : 'a t -> bool

Emptiness test

val map : ( 'a -> 'b ) -> 'a t -> 'b t

map f s maps function f over the stream s

val mapi : ( int -> 'a -> 'b ) -> 'a t -> 'b t

Same as map, but the function is applied to the index of the element as first argument * (counting from 0), and the element itself as second argument.

val iter : ( 'a -> unit ) -> 'a t -> unit

iter f s iterates function f over the stream s

val filter : ( 'a -> bool ) -> 'a t -> 'a t

filter p s filters the stream s using the predicate p * (leaves only those elements x, for which p x = true)

val fold : ( 'a -> 'b -> 'a ) -> 'a -> 'b t -> 'a

fold f a s left-fold over a stream s * with function f and initial accumulator value a

val zip : 'a t -> 'b t -> ('a * 'b) t

zip s s' zips streams s and s' into a stream of pairs; * fails with the Invalid_argument for the streams of different lengths

val mplus : 'a t -> 'a t -> 'a t

mplus s s' monadic-alternative for streams; * concatenates two streams, the resulting stream contains elements * of both input streams in an interleaved order

val bind : 'a t -> ( 'a -> 'b t ) -> 'b t

bind s f monadic-bind for streams; * maps function f over values of the stream s, * obtaining a stream of streams 'b t t, and then flattens this stream

val retrieve : ?n:int -> 'a t -> 'a list * 'a t

retrieve ~n:n s returns the list of n-first elements of s and the rest of the stream

val take : ?n:int -> 'a t -> 'a list

take ~n:n s returns the list of n-first elements of s

val hd : 'a t -> 'a

hd s gets a head of the stream

val tl : 'a t -> 'a t

tl s gets a tail of the stream