Class type OCanren.reified

Reification result.

method is_open : bool

Returns true if the term has any free logic variable inside.

method reify : 'b. ( 'a ilogic, 'b ) Reifier.t -> 'b

Gets the answer as a logic value using provided reifier.