Module OCanren__Core.Tabling

Tabling primitives. Tabling allows to cache answers of the goal between different queries.

Usage: General form : Tabling.tabled/tabledrec n g where n is the number of parameters and g is the goal. Returns modified `tabled` goal.

1) For non-recursive goals:

let g = Tabling.(tabled two) (fun q r -> q === r)

2) For recursive goals: In this case it is necessery to `abstract` from recursive calls. The goal should take additional (first) argument grec and use it instead of recursive calls to itself.

let g = Tabling.(tabledrec one) (fun grec q -> (q === O) ||| (fresh (n) (q === S n) &&& (grec n)))

val succ : (unit -> (('a -> 'b) -> 'c) * ('d -> 'e -> 'f)) -> unit -> (((('g, 'h) OCanren__.Logic.injected * 'a) -> 'b) -> ('g, 'h) OCanren__.Logic.injected -> 'c) * (('i -> 'd) -> ('i * 'e) -> 'f)

val one : unit -> ((('a, 'b) OCanren__.Logic.injected -> 'c) -> ('a, 'b) OCanren__.Logic.injected -> 'c) * (('d -> 'e) -> 'd -> 'e)

val two : unit -> (((('a, 'b) OCanren__.Logic.injected * ('c, 'd) OCanren__.Logic.injected) -> 'e) -> ('a, 'b) OCanren__.Logic.injected -> ('c, 'd) OCanren__.Logic.injected -> 'e) * (('f -> 'g -> 'h) -> ('f * 'g) -> 'h)

val three : unit -> (((('a, 'b) OCanren__.Logic.injected * (('c, 'd) OCanren__.Logic.injected * ('e, 'f) OCanren__.Logic.injected)) -> 'g) -> ('a, 'b) OCanren__.Logic.injected -> ('c, 'd) OCanren__.Logic.injected -> ('e, 'f) OCanren__.Logic.injected -> 'g) * (('h -> 'i -> 'j -> 'k) -> ('h * ('i * 'j)) -> 'k)

val four : unit -> (((('a, 'b) OCanren__.Logic.injected * (('c, 'd) OCanren__.Logic.injected * (('e, 'f) OCanren__.Logic.injected * ('g, 'h) OCanren__.Logic.injected))) -> 'i) -> ('a, 'b) OCanren__.Logic.injected -> ('c, 'd) OCanren__.Logic.injected -> ('e, 'f) OCanren__.Logic.injected -> ('g, 'h) OCanren__.Logic.injected -> 'i) * (('j -> 'k -> 'l -> 'm -> 'n) -> ('j * ('k * ('l * 'm))) -> 'n)

val five : unit -> (((('a, 'b) OCanren__.Logic.injected * (('c, 'd) OCanren__.Logic.injected * (('e, 'f) OCanren__.Logic.injected * (('g, 'h) OCanren__.Logic.injected * ('i, 'j) OCanren__.Logic.injected)))) -> 'k) -> ('a, 'b) OCanren__.Logic.injected -> ('c, 'd) OCanren__.Logic.injected -> ('e, 'f) OCanren__.Logic.injected -> ('g, 'h) OCanren__.Logic.injected -> ('i, 'j) OCanren__.Logic.injected -> 'k) * (('l -> 'm -> 'n -> 'o -> 'p -> 'q) -> ('l * ('m * ('n * ('o * 'p)))) -> 'q)

val tabled : (unit -> (('a -> State.t OCanren.Stream.t goal') -> 'b) * ('c -> 'a -> State.t OCanren.Stream.t goal')) -> 'c -> 'b

val tabledrec : (unit -> (('a -> State.t OCanren.Stream.t goal') -> 'b -> 'c) * ('d -> 'a -> State.t OCanren.Stream.t goal')) -> (('b -> 'c) -> 'd) -> 'b -> 'c