Module Pa_ocanren

Camlp5 syntax extension for miniKanren syntax constructs

There are two syntax extensions provided: fresh and defer.

fresh is a direct analog to the corresponding construct in original miniKanren. It has the form fresh (x y ...) g, where x y ... is a list of free variables, created by fresh, g --- some goal.

defer performs "inverse-eta-delay". It has the form defer (g) and expanded into fun st -> Lazy.from_fun (fun () -> g st).

val fold_right1 : ( 'a -> 'a -> 'a ) -> 'a list -> 'a

val fold_left1 : ( 'a -> 'a -> 'a ) -> 'a list -> 'a

val decapitalize : string -> string

val ctor : MLast.expr -> MLast.expr option

val list_of_list : MLast.expr list -> MLast.expr

val fix_term : MLast.expr -> MLast.expr

val is_operator : string -> bool

val operator_rparen : string Grammar.Entry.e

val operator : string Grammar.Entry.e

val symbolchar : string -> int -> bool

val prefix : string Grammar.Entry.e

val op_from_list : string list -> string

val of_val : 'a Ploc.vala -> 'a

val decorate_type : MLast.ctyp -> MLast.ctyp