Module Relude_Array_Base

let cons: 'a => array('a) => array('a);

* Prepends a single item to the start of an array.

let prepend: 'a => array('a) => array('a);

* Prepends a single item to the start of an array. * Alias for `cons`

let uncons: array('a) => option(('a, array('a)));

* Splits an array into head and tail parts, but only if the array is non-empty.

let append: 'a => array('a) => array('a);

* Appends a single item to the end of an array.

let repeat: int => 'a => array('a);

* Creates an array by repeating the given item the given number of times.

let makeWithIndex: int => (int => 'a) => array('a);

* Creates an array by mapping a function over a range of values from `0` to `n - 1`.

let mapWithIndex: a b. ('a => int => 'b) => array('a) => array('b);

* Maps a pure function over the array which provides the index of the item along with the item itself.

let reverse: array('a) => array('a);

* Reverses an array

let shuffleInPlace: array('a) => array('a);

* Shuffles an array in-place (with mutation)

let shuffle: array('a) => array('a);

* Creates a new array which is a shuffled version of the input array.

let length: array('a) => int;

* Gets the length of the array.

let isEmpty: array('a) => bool;

* Indicates if the array is empty (length 0).

let isNotEmpty: array('a) => bool;

* Indicates if the array is non-empty (length > 0).

let at: int => array('a) => option('a);

* Gets an item at the given index, or returns None.

let setAt: int => 'a => array('a) => option(array('a));

* Sets a value in-place (mutation) at the given index, but only if the index is within * the current range of the array.

let head: array('a) => option('a);

* Gets the first item in the array (if non-empty).

let tail: array('a) => option(array('a));

* Gets all but the the first item in the array (if non-empty).

let tailOrEmpty: array('a) => array('a);

* Gets all but the first item item in the array (if non-empty), or an empty array.

let init: array('a) => option(array('a));

* Gets all but the last item in the array (if non-empty).

let initOrEmpty: array('a) => array('a);

* Gets all but the last item in the array (if non-empty), or the empty array.

let last: array('a) => option('a);

* Gets the last item in the array (if non-empty);

let take: int => array('a) => array('a);

* Gets up to the given count of items from the array. * * If the count is greater than the length, it gets all of the items.

let takeExactly: int => array('a) => option(array('a));

* Gets exactly the given count of items from the array. * * If the count is greater than the length of the array, None is returned.

let takeWhile: ('a => bool) => array('a) => array('a);

* Gets items from the array until it reaches an item which does not satisfy the given predicate.

let drop: int => array('a) => array('a);

* Removes up to the count items from the array. * * If the count is greater than the length of the array, an empty array is returned.

let dropExactly: int => array('a) => option(array('a));

* Removes exactly the count items from the array. * * If the count is greater than the length of the array, None is returned.

let dropWhile: ('a => bool) => array('a) => array('a);

* Drops items from the array until an item is reached which does not satisfy the given predicate.

let filter: a. ('a => bool) => array('a) => array('a);

* Returns a new array of items from the input array which satisfy the given predicate.

let keep: a. ('a => bool) => array('a) => array('a);

* Alias of filter

let filterWithIndex: a. ('a => int => bool) => array('a) => array('a);

* Returns a new array of items from the input array which satisfy the given indexed predicate.

let keepWithIndex: a. ('a => int => bool) => array('a) => array('a);

* Alias of filterWithIndex

let filterNot: a. ('a => bool) => array('a) => array('a);

* Returns a new array of items from the input array which do not satisfy the given predicate.

let reject: a. ('a => bool) => array('a) => array('a);

* Alias of filterNot

let filterNotWithIndex: a. ('a => int => bool) => array('a) => array('a);

* Returns a new array of items from the input array which do not satisfy the given indexed predicate.

let rejectWithIndex: a. ('a => int => bool) => array('a) => array('a);

* Alias of filterNotWithIndex

let mapOption: a b. ('a => option('b)) => array('a) => array('b);

* Maps a function over the array and keeps only the values which were Some.

let catOption: a. array(option('a)) => array('a);

* Collapses an array of options into an array of pure values, keeping only the `Some` values.

let partition: ('a => bool) => array('a) => (array('a), array('a));

* Splits an array into two separate arrays - one containing items which satisfy the predicate, and * the other array containing the items which do not satisfy the predicate. * * The items which satisfy the predicate are on the left side of the resulting tuple.

let splitAt: int => array('a) => option((array('a), array('a)));

* Splits an array into two parts at the given index. * * Returns `None` if the index is out of range.

let prependToAll: 'a => array('a) => array('a);

* Creates a new array which contains the given delimiter inserted before every item in * the input array.

let intersperse: 'a => array('a) => array('a);

* Creates a new array which contains the given delimiter inserted between every item in * the input array.

let replicate: int => array('a) => array('a);

* Creates a new array by concatenating the given array n times.

let zip: array('a) => array('b) => array(('a, 'b));

* Creates a new array by combining pair-wise values from the two input arrays as a tuple.

let zipWith: ('a => 'b => 'c) => array('a) => array('b) => array('c);

* Creates a new array by combining pair-wise values from the two input arrays using the given function.

let zipWithIndex: array('a) => array(('a, int));

* Creates a new array of each value paired with its index in a tuple.

let unzip: array(('a, 'b)) => (array('a), array('b));

* Converts an array of tuples into two separate arrays, one array containing the left values, * and the other array containing the right values.

let sortWithInt: ('a => 'a => int) => array('a) => array('a);

* Sorts an array using an int-based compare function.

let sortBy: ('a => 'a => BsBastet.Interface.ordering) => array('a) => array('a);

* Sorts an array using a compare function.

let sort: (module BsBastet.Interface.ORD with type t = 'a) => array('a) => array('a);

* Sorts an array using an ORD module.

let distinctBy: a. ('a => 'a => bool) => array('a) => array('a);

* Creates a new array containing only the distinct values of the array, using the given * equality function.

let removeFirstBy: a. ('a => 'a => bool) => 'a => array('a) => array('a);

* Removes the first occurrence of the given value from the array, using the given equality function.

let removeEachBy: a. ('a => 'a => bool) => 'a => array('a) => array('a);

* Removes all occurrences of the given value from the array, using the given equality function.

let distinct: (module BsBastet.Interface.EQ with type t = 'a) => array('a) => array('a);

* Creates a new array with only the distinct values from the array, using the given EQ module.

let removeFirst: (module BsBastet.Interface.EQ with type t = 'a) => 'a => array('a) => array('a);

* Removes the first occurrence of the given value from the array, using the given EQ module

let removeEach: (module BsBastet.Interface.EQ with type t = 'a) => 'a => array('a) => array('a);

* Removes all occurrences of the given value from the array, using the given EQ module

let replaceAt: a. int => 'a => array('a) => array('a);

* Creates a new array by replace the value at the given index with the given value. If the * index is out of bounds, no replacement is made.

let scanLeft: ('b => 'a => 'b) => 'b => array('a) => array('b);

* Folds an array from left-to-right, collecting the results of the * accumulation for each individual iteration.

let scanRight: ('a => 'b => 'b) => 'b => array('a) => array('b);

* Folds an array from right-to-left, collecting the results of the * accumulation for each individual iteration.

let insertAt: a. int => 'a => array('a) => array('a);

* Creates a new array that inserts the given value at the given index. * If the index is out of range, no insertion is made.

let updateAt: a. int => ('a => 'a) => array('a) => array('a);

* Creates a new array that modifies the value at the given index with the given function. * If the index is out of range, no change is made.

let swapAt: a. int => int => array('a) => array('a);

* Creates a new array with the elements at the two given indexes swapped. * If either index is out of range, no change is made.

let removeAt: a. int => array('a) => array('a);

* Creates a new array without the element at the given index. * If the index is out of range, no change is made.

let chunk: a. int => array('a) => array(array('a));

* Creates an array of elements split into groups the length of size. * If the array can't be split evenly, the final chunk will be the remaining elements.