Module Relude_Array_Instances

let concat: a. array('a) => array('a) => array('a);

Concatenates two arrays with the left side array first, and the right side last

module SemigroupAny: BsBastet.Interface.SEMIGROUP_ANY with type SemigroupAny.t('a) = array('a);

include { ... };

let concatNamed: prefix:SemigroupAny.t('a) => SemigroupAny.t('a) => SemigroupAny.t('a);

let map: a b. ('a => 'b) => array('a) => array('b);

Applies a pure function to each value in the array

module Functor: BsBastet.Interface.FUNCTOR with type Functor.t('a) = array('a);

include { ... };

module BsFunctorExtensions: { ... };

let flipMap: Functor.t('a) => ('a => 'b) => Functor.t('b);

let void: Functor.t('a) => Functor.t(unit);

let voidRight: 'a => Functor.t('b) => Functor.t('a);

let voidLeft: Functor.t('a) => 'b => Functor.t('b);

let flap: Functor.t(('a => 'b)) => 'a => Functor.t('b);

let apply: a b. array(('a => 'b)) => array('a) => array('b);

Applies an array of functions to an array of values to produce a new array of values.

module Apply: BsBastet.Interface.APPLY with type Apply.t('a) = array('a);

include { ... };

module BsApplyExtensions: { ... };

let applyFirst: Apply.t('a) => Apply.t('b) => Apply.t('a);

let applySecond: Apply.t('a) => Apply.t('b) => Apply.t('b);

let map2: ('a => 'b => 'c) => Apply.t('a) => Apply.t('b) => Apply.t('c);

let map3: ('a => 'b => 'c => 'd) => Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d);

let map4: ('a => 'b => 'c => 'd => 'e) => Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d) => Apply.t('e);

let map5: ('a => 'b => 'c => 'd => 'e => 'f) => Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d) => Apply.t('e) => Apply.t('f);

let tuple2: Apply.t('a) => Apply.t('b) => Apply.t(('a, 'b));

let tuple3: Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t(('a, 'b, 'c));

let tuple4: Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d) => Apply.t(('a, 'b, 'c, 'd));

let tuple5: Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d) => Apply.t('e) => Apply.t(('a, 'b, 'c, 'd, 'e));

let mapTuple2: ('a => 'b => 'c) => (Apply.t('a), Apply.t('b)) => Apply.t('c);

let mapTuple3: ('a => 'b => 'c => 'd) => (Apply.t('a), Apply.t('b), Apply.t('c)) => Apply.t('d);

let mapTuple4: ('a => 'b => 'c => 'd => 'e) => (Apply.t('a), Apply.t('b), Apply.t('c), Apply.t('d)) => Apply.t('e);

let mapTuple5: ('a => 'b => 'c => 'd => 'e => 'f) => (Apply.t('a), Apply.t('b), Apply.t('c), Apply.t('d), Apply.t('e)) => Apply.t('f);

let pure: a. 'a => array('a);

Returns a new array containing the single given item

module Applicative: BsBastet.Interface.APPLICATIVE with type Applicative.t('a) = array('a);

include { ... };

module BsApplicativeExtensions: { ... };

let liftA1: ('a => 'b) => Applicative.t('a) => Applicative.t('b);

let when_: bool => Applicative.t(unit) => Applicative.t(unit);

let unless: bool => Applicative.t(unit) => Applicative.t(unit);

let all: list(Applicative.t('a)) => Applicative.t(list('a));

let bind: a b. array('a) => ('a => array('b)) => array('b);

Maps a monadic function over each element of the array, and flattens (concatenates) the result.

module Monad: BsBastet.Interface.MONAD with type Monad.t('a) = array('a);

include { ... };

module BsMonadExtensions: { ... };

let flatMap: ('a => Monad.t('b)) => Monad.t('a) => Monad.t('b);

let flatten: Monad.t(Monad.t('a)) => Monad.t('a);

let composeKleisli: ('a => Monad.t('b)) => ('b => Monad.t('c)) => 'a => Monad.t('c);

let flipComposeKleisli: ('b => Monad.t('c)) => ('a => Monad.t('b)) => 'a => Monad.t('c);

let liftM1: ('a => 'b) => Monad.t('a) => Monad.t('b);

let when_: Monad.t(bool) => Monad.t(unit) => Monad.t(unit);

let unless: Monad.t(bool) => Monad.t(unit) => Monad.t(unit);

let alt: a. array('a) => array('a) => array('a);

Alt for arrays concatenates the two arrays

module Alt: BsBastet.Interface.ALT with type Alt.t('a) = array('a);

include { ... };

let orElse: fallback:Alt.t('a) => Alt.t('a) => Alt.t('a);

let imap: a b. ('a => 'b) => ('b => 'a) => array('a) => array('b);

Imap is the invariant map function for arrays.

module Invariant: BsBastet.Interface.INVARIANT with type Invariant.t('a) = array('a);

let extend: a b. (array('a) => 'b) => array('a) => array('b);

Extend is the dual of the monadic bind function.

module Extend: BsBastet.Interface.EXTEND with type Extend.t('a) = array('a);

let foldLeft: ('a => 'b => 'a) => 'a => BsBastet.Array.Foldable.t('b) => 'a;

Folds an array from left to right into an accumulator value

let foldRight: ('a => 'b => 'b) => 'b => BsBastet.Array.Foldable.t('a) => 'b;

Folds an array from right-to-left into an accumulator value

module Foldable: BsBastet.Interface.FOLDABLE with type Foldable.t('a) = array('a);

include { ... };

module BsFoldableExtensions: { ... };

let any: ('a => bool) => Foldable.t('a) => bool;

let all: ('a => bool) => Foldable.t('a) => bool;

let containsBy: ('a => 'a => bool) => 'a => Foldable.t('a) => bool;

let contains: (module BsBastet.Interface.EQ with type t = 'a) => 'a => Foldable.t('a) => bool;

let indexOfBy: ('a => 'a => bool) => 'a => Foldable.t('a) => option(int);

let indexOf: (module BsBastet.Interface.EQ with type t = 'a) => 'a => Foldable.t('a) => option(int);

let minBy: ('a => 'a => BsBastet.Interface.ordering) => Foldable.t('a) => option('a);

let min: (module BsBastet.Interface.ORD with type t = 'a) => Foldable.t('a) => option('a);

let maxBy: ('a => 'a => BsBastet.Interface.ordering) => Foldable.t('a) => option('a);

let max: (module BsBastet.Interface.ORD with type t = 'a) => Foldable.t('a) => option('a);

let countBy: ('a => bool) => Foldable.t('a) => int;

let length: Foldable.t('a) => int;

let size: Foldable.t('a) => int;

let count: Foldable.t('a) => int;

let forEach: ('a => unit) => Foldable.t('a) => unit;

let forEachWithIndex: ('a => int => unit) => Foldable.t('a) => unit;

let find: ('a => bool) => Foldable.t('a) => option('a);

let findWithIndex: ('a => int => bool) => Foldable.t('a) => option('a);

let toList: Foldable.t('a) => list('a);

let toArray: Foldable.t('a) => array('a);

module FoldableSemigroupExtensions: (S: BsBastet.Interface.SEMIGROUP) => { ... };

module FoldableMonoidExtensions: (M: BsBastet.Interface.MONOID) => { ... };

let foldMap: (module BsBastet.Interface.MONOID with type t = 'a) => ('b => 'a) => Foldable.t('b) => 'a;

let foldWithMonoid: (module BsBastet.Interface.MONOID with type t = 'a) => Foldable.t('a) => 'a;

let intercalate: (module BsBastet.Interface.MONOID with type t = 'a) => 'a => Foldable.t('a) => 'a;

module FoldableApplicativeExtensions: (A: BsBastet.Interface.APPLICATIVE) => { ... };

module FoldableMonadExtensions: (M: BsBastet.Interface.MONAD) => { ... };

module FoldableEqExtensions: (E: BsBastet.Interface.EQ) => { ... };

module FoldableOrdExtensions: (O: BsBastet.Interface.ORD) => { ... };

module Traversable = BsBastet.Array.Traversable;

let eqBy: a. ('a => 'a => bool) => array('a) => array('a) => bool;

Indicates if two arrays are pair-wise equal, using the given equality function

let eq: (module BsBastet.Interface.EQ with type t = 'a) => array('a) => array('a) => bool;

Indicates if two arrays are pair-wise equal, using the given EQ module

module Eq: (EqA: BsBastet.Interface.EQ) => { ... };

module Ord = BsBastet.Array.Ord;

let showBy: a. ('a => BsBastet.String.Monoid.t) => array('a) => string;

Converts an array to a string, using the given show function for converting the array items

let show: (module BsBastet.Interface.SHOW with type t = 'a) => array('a) => string;

Converts an array to a string, using the given SHOW module for converting the array items

module Show: (ShowA: BsBastet.Interface.SHOW) => { ... };

let fromList: a. list('a) => array('a);

Converts the given list to an array

let toList: a. array('a) => list('a);

Converts the given array to a list

module IsoList: Relude_Interface.ISO_LIST with type IsoList.t('a) = array('a);