Module Relude_Array_Specializations.Float

* Array extensions for array(float)

include { ... };

let contains: Relude_Float.Ord.t => Relude_Array_Instances.Foldable.t(Relude_Float.Ord.t) => bool;

let indexOf: Relude_Float.Ord.t => Relude_Array_Instances.Foldable.t(Relude_Float.Ord.t) => option(int);

let distinct: array(Relude_Float.Ord.t) => array(Relude_Float.Ord.t);

let removeFirst: Relude_Float.Ord.t => array(Relude_Float.Ord.t) => array(Relude_Float.Ord.t);

let removeEach: Relude_Float.Ord.t => array(Relude_Float.Ord.t) => array(Relude_Float.Ord.t);

let eq: array(Relude_Float.Ord.t) => array(Relude_Float.Ord.t) => bool;

let min: Relude_Array_Instances.Foldable.t(Relude_Float.Ord.t) => option(Relude_Float.Ord.t);

let max: Relude_Array_Instances.Foldable.t(Relude_Float.Ord.t) => option(Relude_Float.Ord.t);

let sort: array(Relude_Float.Ord.t) => array(Relude_Float.Ord.t);

let sum: Relude_Array_Instances.Foldable.t(Relude_Float.Additive.Monoid.t) => Relude_Float.Additive.Monoid.t;

* Finds the sum of all the floats in the array

let product: Relude_Array_Instances.Foldable.t(Relude_Float.Multiplicative.Monoid.t) => Relude_Float.Multiplicative.Monoid.t;

* Finds the product of all the floats in the array