Module Relude_Array_Specializations.String

Array extensions for array(string)

include { ... };

let contains: Relude_String.Ord.t => Relude_Array_Instances.Foldable.t(Relude_String.Ord.t) => bool;

let indexOf: Relude_String.Ord.t => Relude_Array_Instances.Foldable.t(Relude_String.Ord.t) => option(int);

let distinct: array(Relude_String.Ord.t) => array(Relude_String.Ord.t);

let removeFirst: Relude_String.Ord.t => array(Relude_String.Ord.t) => array(Relude_String.Ord.t);

let removeEach: Relude_String.Ord.t => array(Relude_String.Ord.t) => array(Relude_String.Ord.t);

let eq: array(Relude_String.Ord.t) => array(Relude_String.Ord.t) => bool;

let min: Relude_Array_Instances.Foldable.t(Relude_String.Ord.t) => option(Relude_String.Ord.t);

let max: Relude_Array_Instances.Foldable.t(Relude_String.Ord.t) => option(Relude_String.Ord.t);

let sort: array(Relude_String.Ord.t) => array(Relude_String.Ord.t);

include { ... };

module BsFoldableMonoidExtensions: { ... };

let foldMap: ('a => Relude_String.Monoid.t) => Relude_Array_Instances.Foldable.t('a) => Relude_String.Monoid.t;

let foldWithMonoid: Relude_Array_Instances.Foldable.t(Relude_String.Monoid.t) => Relude_String.Monoid.t;

let intercalate: Relude_String.Monoid.t => Relude_Array_Instances.Foldable.t(Relude_String.Monoid.t) => Relude_String.Monoid.t;

let join: array(string) => string;

Concatenates an array of strings using an empty string ("") delimiter.

Alias of "foldWithMonoid" for array(string) with "" as the empty value and ++ as the append function.

let joinWith: string => array(string) => string;

Joins an array of strings using the given delimiter string.

Alias of "intercalate" for array(string)

let distinct: BsBastet.Array.Foldable.t(Js.Dict.key) => array(string);

Specialized distinct function that removes duplicate strings in O(n) time by using Js.Dict.