Module Relude_Extensions_Applicative.ApplicativeExtensions

* Extensions for any APPLICATIVE

Parameters

A: BsBastet.Interface.APPLICATIVE

Signature

module BsApplicativeExtensions: { ... };

let liftA1: a b. ('a => 'b) => A.t('a) => A.t('b);

* Lifts a pure function `'a => 'b` into an applicative context `A.t('a) => A.t('b)`

let when_: bool => A.t(unit) => A.t(unit);

* Runs the given applicative effect when the condition is true. If the condition * is false, the effect is not run.

let unless: bool => A.t(unit) => A.t(unit);

* Runs the given applicative effect when the condition is false. If the condition * is true, the effect is not run.

let all: a. list(A.t('a)) => A.t(list('a));