Module Relude_Extensions_Apply.ApplyExtensions

* Extensions for any APPLY

Parameters

A: BsBastet.Interface.APPLY

Signature

module BsApplyExtensions: { ... };

let applyFirst: a b. A.t('a) => A.t('b) => A.t('a);

* Runs the two applicative effects, but only keeps the result from the left side.

let applySecond: a b. A.t('a) => A.t('b) => A.t('b);

* Runs the two applicative effects, but only keeps the result from the right side.

let map2: a b c. ('a => 'b => 'c) => A.t('a) => A.t('b) => A.t('c);

* Runs the applicative effects and combines the results using the given pure function.

let map3: a b c d. ('a => 'b => 'c => 'd) => A.t('a) => A.t('b) => A.t('c) => A.t('d);

* Runs the applicative effects and combines the results using the given pure function.

let map4: a b c d e. ('a => 'b => 'c => 'd => 'e) => A.t('a) => A.t('b) => A.t('c) => A.t('d) => A.t('e);

* Runs the applicative effects and combines the results using the given pure function.

let map5: a b c d e f. ('a => 'b => 'c => 'd => 'e => 'f) => A.t('a) => A.t('b) => A.t('c) => A.t('d) => A.t('e) => A.t('f);

* Runs the applicative effects and combines the results using the given pure function.

let tuple2: a b. A.t('a) => A.t('b) => A.t(('a, 'b));

* Runs the applicative effects and combines the results into a tuple.

let tuple3: a b c. A.t('a) => A.t('b) => A.t('c) => A.t(('a, 'b, 'c));

* Runs the applicative effects and combines the results into a tuple.

let tuple4: a b c d. A.t('a) => A.t('b) => A.t('c) => A.t('d) => A.t(('a, 'b, 'c, 'd));

* Runs the applicative effects and combines the results into a tuple.

let tuple5: a b c d e. A.t('a) => A.t('b) => A.t('c) => A.t('d) => A.t('e) => A.t(('a, 'b, 'c, 'd, 'e));

* Runs the applicative effects and combines the results into a tuple.

let mapTuple2: a b c. ('a => 'b => 'c) => (A.t('a), A.t('b)) => A.t('c);

* Runs the applicative effects in the tuple, and combines the results using the given pure function.

let mapTuple3: a b c d. ('a => 'b => 'c => 'd) => (A.t('a), A.t('b), A.t('c)) => A.t('d);

* Runs the applicative effects in the tuple, and combines the results using the given pure function.

let mapTuple4: a b c d e. ('a => 'b => 'c => 'd => 'e) => (A.t('a), A.t('b), A.t('c), A.t('d)) => A.t('e);

* Runs the applicative effects in the tuple, and combines the results using the given pure function.

let mapTuple5: a b c d e f. ('a => 'b => 'c => 'd => 'e => 'f) => (A.t('a), A.t('b), A.t('c), A.t('d), A.t('e)) => A.t('f);

* Runs the applicative effects in the tuple, and combines the results using the given pure function.