Module Relude_Extensions_Apply.ApplyInfix

Infix operator extensions for any APPLY

Parameters

A: BsBastet.Interface.APPLY

Signature

module ApplyExtensions: { ... };

let (<*>): A.t(('a => 'b)) => A.t('a) => A.t('b);

Applies the applicative function on the left to the applicative value on the right.

let (<*): A.t('a) => A.t('b) => A.t('a);

Runs the two applicative effects, and returns the result from the left side.

let (*>): A.t('a) => A.t('b) => A.t('b);

Runs the two applicative effects, and returns the result from the right side.