Module Relude_Extensions_Bifunctor.BifunctorExtensions

Extensions for any Bifunctor

Parameters

B: BsBastet.Interface.BIFUNCTOR

Signature

let mapLeft: a b c. ('a => 'c) => B.t('a, 'b) => B.t('c, 'b);

Maps a function over the left-side type.

let mapRight: a b d. ('b => 'd) => B.t('a, 'b) => B.t('a, 'd);

Maps a function over the right-side type.

let mapError: ('a => 'b) => B.t('c, 'a) => B.t('c, 'b);

Alias for mapRight

Note: this function only makes sense if your error type is on the right of the Bifunctor, like result('a, 'e), or IO.t('a, 'e).