Module Relude_Extensions_BoundedEnum.BoundedEnumExtensions

Parameters

E: Relude_Interface.BOUNDED_ENUM

Signature

let fromThenToAsList: start:E.t => next:E.t => finish:E.t => list(E.t);

Generates a list of enum values starting at ~start, then going to ~next, then using that step size to generate the rest of the list.