Module Relude_Extensions_Enum.EnumExtensions

Parameters

E: Relude_Interface.ENUM

Signature

let fromToAsList: start:E.t => finish:E.t => list(E.t);

* Generates a list of values of the Enum between the given start and finish values (inclusive)

let upFromAsList: E.t => list(E.t);

* Generates a list of enum values that come after the given start value (not including the start) * Warning: the resulting list could be unbounded here

let upFromIncludingAsList: E.t => list(E.t);

* Generates a list of enum values that come after the given start value (including the start) * Warning: the resulting list could be unbounded here

let downFromAsList: E.t => list(E.t);

* Generates a list of enum values that come before the given start value (not including the start) * Warning: the resulting list could be unbounded here

let downFromIncludingAsList: E.t => list(E.t);

* Generates a list of enum values that come before the given start value (including the start) * Warning: the resulting list could be unbounded here