Module Relude_Extensions_Eq.EqExtensions

* Extensions for any EQ

Parameters

Eq: BsBastet.Interface.EQ

Signature

let eqWithConversion: b. ('b => Eq.t) => eq('b);

* Creates a new equality function by contramapping the given conversion function

let notEq: Relude_Eq.eq(Eq.t);

* Indicates if the two items are not equal

let eqInverted: Relude_Eq.eq(Eq.t);

* Alias for notEq

module EqInverted: BsBastet.Interface.EQ with type EqInverted.t = Eq.t;

* An Eq module which is the inverse of the given Eq module

module type EQ_BY_F = (A: Relude_Interface.FUNCTION_1 with type EQ_BY_F.b = Eq.t) => BsBastet.Interface.EQ with type EQ_BY_F.t = A.a;

module EqBy: EQ_BY_F;

* Creates a new Eq for a type b, given an Eq for type a and a Arrow for `b => a` * * Example: * ``` * // Create an Eq for User, given an Eq for string and a function from User => string * module UserEq = String.EqBy( * type a = User.t, * type b = string; * let f = user => user.email; * ); * ```