Module EqInfix.EqExtensions

let eqWithConversion: ('b => Eq.t) => Relude_Eq.eq('b);

let notEq: Relude_Eq.eq(Eq.t);

let eqInverted: Relude_Eq.eq(Eq.t);

module EqInverted: { ... };

An Eq module which is the inverse of the given Eq module

module type EQ_BY_F = (A: { ... }) => { ... };

module EqBy: (A: { ... }) => { ... };

Creates a new Eq for a type b, given an Eq for type a and a function b => a.