Module Relude_Extensions_Foldable.FoldableExtensions

* Extensions for any FOLDABLE

Parameters

F: BsBastet.Interface.FOLDABLE

Signature

module BsFoldableExtensions: { ... };

let any: a. ('a => bool) => F.t('a) => bool;

* Indicates if any item in the foldable satisfies the given predicate

let all: a. ('a => bool) => F.t('a) => bool;

* Indicates if all items in the foldable satisfy the given predicate

let containsBy: a. ('a => 'a => bool) => 'a => F.t('a) => bool;

* Indicates if the foldable contains the given item, using the given equality function

let contains: (module BsBastet.Interface.EQ with type t = 'a) => 'a => F.t('a) => bool;

* Indicates if the foldable contains the given item, using the given EQ module

let indexOfBy: a. ('a => 'a => bool) => 'a => F.t('a) => option(int);

* Finds the index of the given item in the foldable using the given equality function. * If the item is not found, the result is None.

let indexOf: (module BsBastet.Interface.EQ with type t = 'a) => 'a => F.t('a) => option(int);

* Finds the index of the given item in the foldable using the given EQ module. * If the item is not found, the result is None.

let minBy: a. ('a => 'a => BsBastet.Interface.ordering) => F.t('a) => option('a);

* Finds the minimum item in the foldable, using the given comparison function.

let min: (module BsBastet.Interface.ORD with type t = 'a) => F.t('a) => option('a);

* Finds the minimum item in the foldable, using the given ORD module.

let maxBy: a. ('a => 'a => BsBastet.Interface.ordering) => F.t('a) => option('a);

* Finds the maximum item in the foldable, using the given comparison function.

let max: (module BsBastet.Interface.ORD with type t = 'a) => F.t('a) => option('a);

* Finds the maximum item in the foldable, using the given ORD module.

let countBy: a. ('a => bool) => F.t('a) => int;

* Counts the number of items in the foldable which satisfy the given predicate.

let length: a. F.t('a) => int;

* Gets the length of the foldable. * * Alias of size/count

let size: a. F.t('a) => int;

* Alias for length/count

let count: a. F.t('a) => int;

* Alias for length/size

let forEach: a. ('a => unit) => F.t('a) => unit;

* Runs a side-effect function for each item in the foldable.

let forEachWithIndex: a. ('a => int => unit) => F.t('a) => unit;

* Runs a side-effect function for each indexed item in the foldable.

let find: a. ('a => bool) => F.t('a) => option('a);

* Finds the first item in the foldable which satisfies the given predicate.

let findWithIndex: a. ('a => int => bool) => F.t('a) => option('a);

* Finds the first indexed item in the foldable which satisfies the given predicate.

let toList: F.t('a) => list('a);

* Converts the foldable into a list

let toArray: F.t('a) => array('a);

* Converts the foldable into an array

module FoldableSemigroupExtensions: (S: BsBastet.Interface.SEMIGROUP) => { ... };

* Foldable extensions for when you have a Semigroup instance

module FoldableMonoidExtensions: (M: BsBastet.Interface.MONOID) => { ... };

* Foldable extensions for when you have a Monoid instance

let foldMap: (module BsBastet.Interface.MONOID with type t = 'a) => ('b => 'a) => F.t('b) => 'a;

* Maps a function over a foldable, and accumulates the results using the given Monoid module.

let foldWithMonoid: (module BsBastet.Interface.MONOID with type t = 'a) => F.t('a) => 'a;

* Folds a foldable of a monoidal type, accumulating the result using the given Monoid module.

let intercalate: (module BsBastet.Interface.MONOID with type t = 'a) => 'a => F.t('a) => 'a;

* Folds a foldable, accumulating values in a monoid, combining adjacent elements using the specified * separator.

module FoldableApplicativeExtensions: (A: BsBastet.Interface.APPLICATIVE) => { ... };

* Foldable extensions for when you additionally have an Applicative instance.

module FoldableMonadExtensions: (M: BsBastet.Interface.MONAD) => { ... };

* Foldable extensions for when you additionally have a MONAD instance.

module FoldableEqExtensions: (E: BsBastet.Interface.EQ) => { ... };

* Foldable extensions for when you additionally have an EQ instance.

module FoldableOrdExtensions: (O: BsBastet.Interface.ORD) => { ... };

* Foldable extensions for when you additionally have an ORD instance.