Module FoldableExtensions.FoldableApplicativeExtensions

* Foldable extensions for when you additionally have an Applicative instance.

Parameters

A: BsBastet.Interface.APPLICATIVE

Signature

module BsFoldableApplicativeExtensions: { ... };

let traverse_: a b. ('a => A.t('b)) => F.t('a) => A.t(unit);

* Traverse which performs the side effects of the traversal, but throws away the final result (produces unit).

let sequence_: a. F.t(A.t('a)) => A.t(unit);

* Sequence which performs the side effects of the sequence, but throws away the final result (produces unit).