Module FoldableExtensions.FoldableMonoidExtensions

* Foldable extensions for when you have a Monoid instance

Parameters

M: BsBastet.Interface.MONOID

Signature

module BsFoldableMonoidExtensions: { ... };

let foldMap: a. ('a => M.t) => F.t('a) => M.t;

* Maps a function over the foldable, and collects the results via the Monoid instance.

let foldWithMonoid: F.t(M.t) => M.t;

* Folds a foldable of a monoidal type by accumlating values via the monoid instance.

let intercalate: M.t => F.t(M.t) => M.t;

* Folds a foldable, accumulating values in a monoid, combining adjacent elements using the specified * separator.