Module FoldableExtensions.FoldableOrdExtensions

Foldable extensions for when you additionally have an ORD instance.

Parameters

O: BsBastet.Interface.ORD

Signature

let min: F.t(O.t) => option(O.t);

Gets the minimum value of the foldable, using the given ORD module.

let max: F.t(O.t) => option(O.t);

Gets the maximum value of the foldable, using the given ORD module.