Module Relude_Extensions_Functor.FunctorExtensions

* Extensions for any FUNCTOR

Parameters

F: BsBastet.Interface.FUNCTOR

Signature

module BsFunctorExtensions: { ... };

let flipMap: a b. F.t('a) => ('a => 'b) => F.t('b);

* Flipped version of the map function which has the functor on the left, and the function on the right.

let void: a. F.t('a) => F.t(unit);

* Clears the value(s) of a functor by mapping a function that produces unit for each value in the functor.

let voidRight: a b. 'a => F.t('b) => F.t('a);

* Replaces the value(s) of the functor on the right with the value on the left.

let voidLeft: a b. F.t('a) => 'b => F.t('b);

* Replaces the value(s) of the functor on the left with the value on the right.

let flap: a b. F.t(('a => 'b)) => 'a => F.t('b);

* Applies an argument of type `'a` to a functor of `'a => 'b` functions, to produce a functor of `'b`