Module Relude_Extensions_Functor.FunctorInfix

Infix operator extensions for any FUNCTOR

Parameters

F: BsBastet.Interface.FUNCTOR

Signature

module FunctorExtensions: { ... };

let (<$>): ('a => 'b) => F.t('a) => F.t('b);

Operator version of the map function, which has the function on the left, and functor on the right.

let (<#>): F.t('a) => ('a => 'b) => F.t('b);

Operator version of the flipMap function, which has the functor on the left, and function on the right.

let (<$): 'a => F.t('b) => F.t('a);

Operator version of voidRight, which replaces the values of the functor on the right with the value on the left.

let ($>): F.t('a) => 'b => F.t('b);

Operator version of voidLeft, which replaces the values of the functor on the left with the value on the right.

let (<@>): F.t(('a => 'b)) => 'a => F.t('b);

Operator version of flap, which takes a functor of functions F.t('a => 'b) on the left, and an 'a on the right, and applies the value 'a to each function, producing a functor of 'b (F.t('b)).