Module Relude_Extensions_Monad

module MonadExtensions: (M: BsBastet.Interface.MONAD) => { ... };

Extensions.Monad contains a module functor MonadExtensions which gives you access to a wide variety of functions and infix operators that you can get "for free" when you have a Monad typeclass instance.

module MonadInfix: (M: BsBastet.Interface.MONAD) => { ... };

Infix operator extensions for MONAD