Module Relude_Extensions_Monad.MonadExtensions

Extensions.Monad contains a module functor MonadExtensions which gives you access to a wide variety of functions and infix operators that you can get "for free" when you have a Monad typeclass instance.

Extensions for any MONAD

Parameters

M: BsBastet.Interface.MONAD

Signature

module BsMonadExtensions: { ... };

let flatMap: a b. ('a => M.t('b)) => M.t('a) => M.t('b);

Flipped version of bind which has the function on the left and the monad on the right. We're calling this flatMap because the signature closely resembles the signature of map with the function on the left, and flatMap is another commonly used name for bind.

let flatten: a. M.t(M.t('a)) => M.t('a);

Flattens a nested monadic structure one time.

let composeKleisli: a b c. ('a => M.t('b)) => ('b => M.t('c)) => 'a => M.t('c);

Creates a new monadic function by composing two monadic functions from left-to-right.

let flipComposeKleisli: a b c. ('b => M.t('c)) => ('a => M.t('b)) => 'a => M.t('c);

Creates a new monadic function by composing two monadic functions from right-to-left.

let liftM1: a b. ('a => 'b) => M.t('a) => M.t('b);

Lifts a pure function 'a => 'b into a monadic context M.t('a) => M.t('b)

let when_: M.t(bool) => M.t(unit) => M.t(unit);

Runs a monadic effect if the given monadic condition is evaluated to true. If the condition is false, a pure unit value is returned with no effect being run.

let unless: M.t(bool) => M.t(unit) => M.t(unit);

Runs a monadic effect if the given monadic condition is evaluated to false. If the condition is true, a pure unit value is returned with no effect being run.