Module Relude_Extensions_Monad.MonadInfix

Infix operator extensions for MONAD

Parameters

M: BsBastet.Interface.MONAD

Signature

module MonadExtensions: { ... };

let (>>=): M.t('a) => ('a => M.t('b)) => M.t('b);

Operator version of applying a monadic function with the monad value on the left, and the function on the right.

let (=<<): ('a => M.t('b)) => M.t('a) => M.t('b);

Operator version of applying a monadic function with the monad value on the right, and the function on the left.

let (>=>): ('a => M.t('b)) => ('b => M.t('c)) => 'a => M.t('c);

Creates a new monadic function by composing two monadic functions from left-to-right.

let (<=<): ('a => M.t('b)) => ('c => M.t('a)) => 'c => M.t('b);

Creates a new monadic function by composing two monadic functions from right-to-left.