Module Relude_Extensions_Monoid.MonoidExtensions

* Extensions for any MONOID

Parameters

M: BsBastet.Interface.MONOID

Signature

module BsMonoidExtensions: { ... };

let guard: bool => M.t => M.t;

* Returns the monoidal value if the given condition is true, otherwise empty.

let power: M.t => int => M.t;

* Combines the given monoidal value with itself the given number of times.