Module Relude_Extensions_Ord

module OrdExtensions: (O: BsBastet.Interface.ORD) => { ... };

Extensions.Ord is a module functor that accepts any ORD-compatible module and returns a collection of extra helper functions specific to that type. For example, imagine we define a custom weekday type that can be compared for ordering:

module OrdInfix: (O: BsBastet.Interface.ORD) => { ... };