Module Relude_Extensions_Ord.OrdExtensions

Extensions.Ord is a module functor that accepts any ORD-compatible module and returns a collection of extra helper functions specific to that type. For example, imagine we define a custom weekday type that can be compared for ordering:

type weekday = Sun | Mon | Tue | Wed | Thu | Fri | Sat;

// simple variants like this can be safely compared with the comparison
// operations from the OCaml stdlib. This comparison uses the order in which
// the constructors were defined, and we convert to Bastet ordering
let compare = (a: weekday, b: weekday) =>
  a < b ? `less_than : a > b ? `greater_than : equal_to;

let eq = (a: weekday, b: weekday) => a == b;

// now we can construct a module of extra Ord functions
module WeekdayOrd = Relude.Extensions.Ord.OrdExtensions({
  type t = weekday;
  let eq = eq;
  let compare = compare;
});

// and we can use these functions with our weekday type
let isBetween = WeekdayOrd.between(~min=Tue, ~max=Fri, Wed); // true

Parameters

O: BsBastet.Interface.ORD

Signature

let compareWithConversion: ('b => O.t) => Relude_Ord.compare('b);

let compareReversed: Relude_Ord.compare(O.t);

module OrdReversed: BsBastet.Interface.ORD with type OrdReversed.t = O.t;

Creates a new Ord module which is the reverse of the given Ord

let compareAsInt: O.t => O.t => int;

let min: O.t => O.t => O.t;

let max: O.t => O.t => O.t;

let lessThan: O.t => O.t => bool;

let lt: O.t => O.t => bool;

let lessThanOrEq: O.t => O.t => bool;

let lte: O.t => O.t => bool;

let greaterThan: O.t => O.t => bool;

let gt: O.t => O.t => bool;

let greaterThanOrEq: O.t => O.t => bool;

let gte: O.t => O.t => bool;

let clamp: min:O.t => max:O.t => O.t => O.t;

let between: min:O.t => max:O.t => O.t => bool;

module OrdRingExtensions: (R: BsBastet.Interface.RING with type OrdRingExtensions.t = O.t) => { ... };

module OrdNamed: { ... };

OrdNamed contains versions of the comparison functions that have a named argument to disambiguate what's on the lhs vs. rhs of the comparison. The ~compareTo values are on the rhs.

module type ORD_BY_F = (A: Relude_Interface.FUNCTION_1 with type ORD_BY_F.b = O.t) => BsBastet.Interface.ORD with type ORD_BY_F.t = A.a;

module OrdBy: ORD_BY_F;

Creates an ORD for type b given this ORD of type a and a function b => a to act as the contravariant.