Module OrdInfix.OrdExtensions

let compareWithConversion: ('b => O.t) => Relude_Ord.compare('b);

let compareReversed: Relude_Ord.compare(O.t);

module OrdReversed: { ... };

Creates a new Ord module which is the reverse of the given Ord

let compareAsInt: O.t => O.t => int;

let min: O.t => O.t => O.t;

let max: O.t => O.t => O.t;

let lessThan: O.t => O.t => bool;

let lt: O.t => O.t => bool;

let lessThanOrEq: O.t => O.t => bool;

let lte: O.t => O.t => bool;

let greaterThan: O.t => O.t => bool;

let gt: O.t => O.t => bool;

let greaterThanOrEq: O.t => O.t => bool;

let gte: O.t => O.t => bool;

let clamp: min:O.t => max:O.t => O.t => O.t;

let between: min:O.t => max:O.t => O.t => bool;

module OrdRingExtensions: (R: { ... }) => { ... };

module OrdNamed: { ... };

OrdNamed contains versions of the comparison functions that have a named argument to disambiguate what's on the lhs vs. rhs of the comparison. The ~compareTo values are on the rhs.

module type ORD_BY_F = (A: { ... }) => { ... };

module OrdBy: (A: { ... }) => { ... };

Creates an ORD for type b given this ORD of type a and a function b => a to act as the contravariant.