Module OrdExtensions.OrdBy

* Creates an ORD for type b given this ORD of type a and an ARROW `b => a` to act as the contravariant * * Example: * ``` * // Create an Ord for a user using an Ord for string and a function from User => string * module UserOrd = String.OrdBy( * type a = User.t; * type b = string; * let f = user => user.email; * ); * ```

Parameters

A: { ... }

Signature

type t = A.a;

let eq: t => t => bool;

let compare: t => t => BsBastet.Interface.ordering;