Module Relude_Float

let eq: float => float => bool;

`Relude.Float` contains typeclass instances and utility functions for the `float` type.

* Indicates if two floats are exactly equal

module Eq: BsBastet.Interface.EQ with type Eq.t = float;

let zero: float;

* 0.0 value

let one: float;

* 1.0 value

let nan: float;

* Value for "not a number", used for mathematical operations that do not result * in a real number, such as `0.0 /. 0.0` and `sqrt(-2)`.

let infinity: float;

* Positive infinity

let negativeInfinity: float;

* Negative infinity

let add: float => float => float;

* Finds the sum of two floats

let subtract: float => float => float;

* Finds the difference of two floats

let multiply: float => float => float;

* Finds the product of two floats

let divide: float => float => float;

* Finds the quotient of two floats

let pow: float => float => float;

* Raise the first number to the power of the second number. * * ```re * pow(2.0, 4.0) == 16.0; * pow(3.0, 2.0) == 9.0; * pow(-2.0, (0.333333)) |> isNaN; * ```

let sqrt: float => float;

* Find the square root of the given float. The square root of negative numbers * is `nan`.

let top: float;

* The maximum float value.

let bottom: float;

* The minimum float value. Note that when using Bucklescript, this value is * hard-coded and is not necessarily equal to `Number.MIN_VALUE` in JS.

let isNaN: float => bool;

* NaN values are never equal to other numeric values, including other NaN * values. This means that the intuitive `myValue == nan` is not sufficient for * determining whether a value is NaN. * * ```re * isNaN(3.14) == false; * isNaN(infinity) == false; * isNaN(nan) == true; * ```

let compare: float => float => BsBastet.Interface.ordering;

* Compates two floats

module Ord: BsBastet.Interface.ORD with type Ord.t = float;

include { ... };

let compareWithConversion: ('b => Ord.t) => Relude_Extensions_Ord.compare('b);

let compareReversed: Relude_Ord.compare(Ord.t);

module OrdReversed: { ... };

let compareAsInt: Ord.t => Ord.t => int;

let min: Ord.t => Ord.t => Ord.t;

let max: Ord.t => Ord.t => Ord.t;

let lessThan: Ord.t => Ord.t => bool;

let lt: Ord.t => Ord.t => bool;

let lessThanOrEq: Ord.t => Ord.t => bool;

let lte: Ord.t => Ord.t => bool;

let greaterThan: Ord.t => Ord.t => bool;

let gt: Ord.t => Ord.t => bool;

let greaterThanOrEq: Ord.t => Ord.t => bool;

let gte: Ord.t => Ord.t => bool;

let clamp: min:Ord.t => max:Ord.t => Ord.t => Ord.t;

let between: min:Ord.t => max:Ord.t => Ord.t => bool;

module OrdRingExtensions: (R: { ... }) => { ... };

module OrdNamed: { ... };

module type ORD_BY_F = (A: { ... }) => { ... };

module OrdBy: (A: { ... }) => { ... };

module Semiring: BsBastet.Interface.SEMIRING with type Semiring.t = float;

include { ... };

module Ring: BsBastet.Interface.RING with type Ring.t = float;

include { ... };

let (-): Ring.t => Ring.t => Ring.t;

let negate: Ring.t => Ring.t;

include { ... };

let abs: Ring.t => Ring.t;

let signum: Ring.t => Ring.t;

module EuclideanRing: BsBastet.Interface.EUCLIDEAN_RING with type EuclideanRing.t = float;

let approximatelyEqual: tolerance:float => float => float => bool;

`approximatelyEqual(~tolerance=t, x, y)` returns `true` if `x` and `y` are within `t` of one another, `false` otherwise. The `~tolerance` must be non-negative.

### Example ```re approximatelyEqual(~tolerance=0.01, 1.500, 1.495) == true; approximatelyEqual(~tolerance=0.01, 1.500, 1.49) == false; ```

let toInt: float => int;

`toInt` converts a float to an int by dropping the fractional part.

Floats that can't be represented as an int (e.g. `infinity`, `nan`) will return 0.

let fromInt: int => float;

`fromInt` converts an int (e.g. 1) to its floating-point representation (1.0)

let fractionalPart: float => Ring.t;

Return only the decimal portion as a positive floating point number, with the whole-number portion set to 0.

Note that the returned value is subject to floating point rounding errors and probably won't be exactly equal to the fractional part of the original number.

### Example ```re fractionalPart(3.141592) ~= 0.141592; fractionalPart(-12.3456) ~= 0.3456; ```

let floor: float => float;

Round a floating point number to the nearest lower whole number.

let floorAsInt: float => int;

Round a floating point number to the nearest lower integer.

let ceil: float => float;

Round a floating point number to the nearest higher whole number.

let ceilAsInt: float => int;

Round a floating point number to the nearest higher integer.

let round: float => float;

Round a floating point number to the nearest whole number.

let roundAsInt: float => int;

Round a floating point number to the nearest integer

let toPrecision: decimals:int => float => float;

`toPrecision` drops decimals so that the given float has no more than the requested number of decimals.

### Example ```re toPrecision(~decimals=2, 3.141592) == 3.14; toPrecision(~decimals=4, -4.99999999999) == -4.9999; ```

let show: float => string;

`toString(x)` returns the string representation of `x`. Note that this may not be in the same form as you wrote it in the source code.

### Example ```re toString(3.49) == "3.49"; toString(7.0000) == "7"; toString(1.0e+03) == "1000"; ```

let toString: float => string;

* Alias for `show`

module Show: BsBastet.Interface.SHOW with type Show.t = float;

let fromString: string => option(float);

`fromString(s)` returns `Some(x)` if the given string represents a valid floating point number, `None` otherwise.

### Example ```re fromString("1000.0") == Some(1000.0); fromString("-5.25E2") == Some(-525.0); fromString("3.4.5") == None; ```

module Additive: { ... };

module Multiplicative: { ... };

module Subtractive: { ... };

module Divisive: { ... };

module Infix: { ... };