Module WithFunctor.WithApplicative

Specifies an applicative into which we will interpret our free monadic program

We also need a Natural Transformation module here, because putting that function inline doesn't seem to work with the existential type captured by the FreeAp (at least I couldn't figure it out).

Parameters

A: BsBastet.Interface.APPLICATIVE

Signature

module WithNT: (NT: Relude_Interface.NATURAL_TRANSFORMATION with type WithNT.f('a) = F.t('a) and type WithNT.g('a) = A.t('a)) => { ... };

We also need a natural transformation in order to create our interpreters for the free applicative. Because of the existential type captured by the free applicative and what appear to be limitations of OCaml, the NT is provided via a module, so that we can achieve the forall a. f a -> g a.