Module WithApplicative.WithNT

We also need a natural transformation in order to create our interpreters for the free applicative. Because of the existential type captured by the free applicative and what appear to be limitations of OCaml, the NT is provided via a module, so that we can achieve the forall a. f a -> g a.

I tried to get this to work using a F.t('a) => A.t('a) function as a normal function argument, but I couldn't figure it out. This worked for free monad, but I think the problem here is the existential "x" type that is captured by this type.

Parameters

NT: Relude_Interface.NATURAL_TRANSFORMATION with type NT.f('a) = F.t('a) and type NT.g('a) = A.t('a)

Signature

let foldFree: a. t('a) => A.t('a);

Applies an interpreter function (natural transformation) to interpret each value of our algebra into a target applicative.