Module WithApplicativeAndNT.A

module WithNT: (NT: { ... }) => { ... };

We also need a natural transformation in order to create our interpreters for the free applicative. Because of the existential type captured by the free applicative and what appear to be limitations of OCaml, the NT is provided via a module, so that we can achieve the forall a. f a -> g a.