Module WithFunctor.WithMonad

* Specifies a monad into which we will interpret our free monadic program

Parameters

M: BsBastet.Interface.MONAD

Signature

let foldFree: (F.t(t('a)) => M.t(t('a))) => t('a) => M.t('a);

* Applies an interpreter function to interpret each value of our algebra into * a target monad.