Module Relude_HMap.Witness

* Witness contains a type t with an existential type variable - a type that we * capture, but lose knowledge of once captured. * `t(_)` is an extensible variant type, indicated by the `..`. This is used below in WITNESS * so that we can add new cases to the Witness.t(_) type on demand, so that we can create witnesses * for any new types we encounter going forward. * https://caml.inria.fr/pub/docs/manual-ocaml/manual037.html

type t(_) = ..;