Module HMAP_TYPE.Key

Key-related types and operations for an HMap

type keyMeta('a);

keyMeta('a) is the metadata we store with the key, like labels, and other functions for operating on the otherwise existential 'a value.

let create: keyMeta('a) => keyImpl('a);

Creates a keyImpl (an actual map key) for the given key metadata

let keyMeta: keyImpl('a) => keyMeta('a);

Extracts the key metadata from a keyImpl (an actual map key)

type t;

The abstract type of our map key

let hideType: keyImpl('a) => t;

Wraps the type-aware keyImpl('a) into the abstract key type

let eq: t => t => bool;

Checks if two keys are equal

let compare: t => t => int;

Compares two keys