Module Relude_Js_Json

type json = Js.Json.t;

Type alias for Js.Json.t.

type dict = Js.Dict.t(json);

Type alias for Js.Dict.t(Js.Json.t)

let show: ?⁠indentSpaces:int => json => string;

Shows a JSON value with optional space indentation

let isNull: json => bool;

Checks if the Js.Json.t value is a null

let isBool: json => bool;

Checks if the Js.Json.t value is a bool

let isString: json => bool;

Checks if the Js.Json.t value is a string

let isNumber: json => bool;

Checks if the Js.Json.t value is a number

let isObject: json => bool;

Checks if the Js.Json.t value is a JSON object

let isArray: json => bool;

Checks if the Js.Json.t value is a JSON array

let null: json;

Creates a Js.Json.t null value

let fromBool: bool => json;

Creates a Js.Json.t boolean value from a boolean

let fromString: string => json;

Creates a Js.Json.t string value from a string

let fromInt: int => json;

Creates a Js.Json.t number value from an int

let fromFloat: float => json;

Creates a Js.Json.t number value from a float

let fromOption: ('a => json) => option('a) => json;

Creates a Js.Json.t value from an option('a). If the option is None, a null value is returned, otherwise, the value is encoded using the given function.

let fromArrayOfJson: array(json) => json;

Creates a Js.Json.t array value from an array of Js.Json.t values

let fromArrayOfJsonBy: a. ('a => json) => array('a) => json;

Creates a Js.Json.t array value from an array of values that can be converted to Js.Json.t values with the given function

let fromListOfJson: list(json) => json;

Creates a Js.Json.t array value from an list of Js.Json.t values

let fromListOfJsonBy: a. ('a => json) => list('a) => json;

Creates a Js.Json.t array value from an list of values that can be converted to Js.Json.t values with the given function

let fromDictOfJson: dict => json;

Creates a Js.Json.t object value from a Js.Dict containing Js.Json.t values.

let fromArrayOfDictOfJson: array(dict) => json;

Creates a Js.Json.t array value from an array of Js.Dict.t(Js.Json.t) values.

let fromListOfDictOfJson: list(dict) => json;

Creates a Js.Json.t array value from an list of Js.Dict.t(Js.Json.t) values.

let fromArrayOfKeyValueTuples: array((Js.Dict.key, json)) => json;

Creates a Js.Json.t object value from an array of key/value (string/Js.Json.t) tuples.

let fromListOfKeyValueTuples: list((Js.Dict.key, json)) => json;

Creates a Js.Json.t object value from an list of key/value (string/Js.Json.t) tuples.

let toNull: json => option(unit);

Attempts to decode the given Js.Json.t value as a null. Returns Some(()) if it is a null, otherwise None.

let toBool: json => option(bool);

Attempts to decode the given Js.Json.t value as a boolean

let toString: json => option(string);

Attempts to decode the given Js.Json.t value as a string

let toInt: json => option(int);

Attempts to decode the given Js.Json.t value as an int

let toFloat: json => option(float);

Attempts to decode the given Js.Json.t value as a float

let toArrayOfJson: json => option(array(json));

Attempts to decode the given Js.Json.t value as an array of Js.Json.t values.

let toArrayOfJsonOrElse: array(json) => json => array(json);

Attempts to decode the given Js.Json.t value as an array of Js.Json.t values, with a fallback.

let toArrayOfJsonOrEmpty: json => array(json);

Attempts to decode the given Js.Json.t value as an array of Js.Json.t values, with a fallback of an empty array.

let toListOfJson: json => option(list(json));

Attempts to decode the given Js.Json.t value as a list of Js.Json.t values.

let toListOfJsonOrElse: list(json) => json => list(json);

Attempts to decode the given Js.Json.t value as an list of Js.Json.t values, with a fallback.

let toListOrEmpty: json => list(json);

Attempts to decode the given Js.Json.t value as a list of Js.Json.t values, with a fallback of an empty list.

let toDictOfJson: json => option(dict);

Attempts to decode the given Js.Json.t value as a Js.Dict.t(Js.Json.t)

let toDictOfJsonOrElse: dict => json => dict;

Attempts to decode the given Js.Json.t value as a Js.Dict.t(Js.Json.t) with a fallback.

let toDictOfJsonOrEmpty: json => dict;

Attempts to decode the given Js.Json.t value as a Js.Dict.t(Js.Json.t) with a fallback of an empty Js.Dict.

module Error: { ... };

module Errors: { ... };

module ValidationE: { ... };

module ArrayValidationE: { ... };

module TraversableE = ArrayValidationE.Traversable;

let validateNull: json => Relude_Validation.t(unit, Errors.t);

Validates that the given Js.Json.t value is a null

let validateBool: json => Relude_Validation.t(bool, Errors.t);

Validates that the given Js.Json.t value is a bool

let validateString: json => Relude_Validation.t(string, Errors.t);

Validates that the given Js.Json.t value is a string

let validateInt: json => Relude_Validation.t(int, Errors.t);

Validates that the given Js.Json.t value is an int

let validateFloat: json => Relude_Validation.t(float, Errors.t);

Validates that the given Js.Json.t value is a float

let validateOptional: ?⁠errorAsNone:bool => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(option('a), Errors.t);

Validates that the given value is either null or can be decoded using the given decoder.

Failed validation can be treated as None or returned as an error using the errorAsNone flag.

let getJsonAtIndex: int => json => option(json);

Validates that the given Js.Json.t value is an array, and attempts to get the value at the given index as a raw Js.Json.t value.

let validateJsonAtIndex: int => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t('a, Errors.t);

Validates that the given Js.Json.t value is an array, and attempts to get the value at the given index and validate it with the given validation function.

let validateNullAtIndex: int => json => Relude_Validation.t(unit, Errors.t);

Validates that the given Js.Json.t value is an array with a null at the given index.

let validateBoolAtIndex: int => json => Relude_Validation.t(bool, Errors.t);

Validates that the given Js.Json.t value is an array with a bool at the given index.

let validateIntAtIndex: int => json => Relude_Validation.t(int, Errors.t);

Validates that the given Js.Json.t value is an array with an int at the given index.

let validateFloatAtIndex: int => json => Relude_Validation.t(float, Errors.t);

Validates that the given Js.Json.t value is an array with a float at the given index.

let validateStringAtIndex: int => json => Relude_Validation.t(string, Errors.t);

Validates that the given Js.Json.t value is an array with a string at the given index.

let validateOptionalAtIndex: ?⁠missingAsNone:bool => ?⁠nullAsNone:bool => ?⁠errorAsNone:bool => int => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(option('a), Errors.t);

Validates the given Js.Json.t value at the given index, using the validation function.

An invalid index can be treated as None or returned as an error using the missingAsNone flag. A null value can be treated as None or returned as an error using the nullAsNone flag. Failed validation can be treated as None or returned as an error using the errorAsNone flag.

let validateArrayOfJson: a. (int => json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(array('a), Errors.t);

Validates that the given Js.Json.t value is an array, then validates each item of the array using the given validation function.

let validateArrayOfJsonAsList: a. (int => json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(list('a), Errors.t);

Validates that the given Js.Json.t value is an array, then validates each item of the array using the given validation function, then converts the result to a list.

let validateArrayAtIndex: a. int => (int => json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(array('a), Errors.t);

Validates that the given Js.Json.t value is an array, and then validates the value at the given index is an array, and validates it using the given validation function.

let validateObjectAtIndex: a. int => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t('a, Errors.t);

Validates that the Js.Json.t value is an array, then validates that the value at the given index is a Json object, and validates the object using the given validation function.

let getJsonForKey: string => json => option(json);

Validates that the given Js.Json.t value is an object, and then gets the raw Js.Json.t value for the given key.

let validateJsonForKey: string => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t('a, Errors.t);

Validates that the given Js.Json.t value is an object, then validates the value at the given key using the given validation function.

let validateNullForKey: string => json => Relude_Validation.t(unit, Errors.t);

Validates the given Js.Json.t value is an object with a null at the given key.

let validateBoolForKey: string => json => Relude_Validation.t(bool, Errors.t);

Validates the given Js.Json.t value is an object with a bool at the given key.

let validateIntForKey: string => json => Relude_Validation.t(int, Errors.t);

Validates the given Js.Json.t value is an object with an int at the given key.

let validateFloatForKey: string => json => Relude_Validation.t(float, Errors.t);

Validates the given Js.Json.t value is an object with a float at the given key.

let validateStringForKey: string => json => Relude_Validation.t(string, Errors.t);

Validates the given Js.Json.t value is an object with a string at the given key.

let validateOptionalForKey: ?⁠missingAsNone:bool => ?⁠nullAsNone:bool => ?⁠errorAsNone:bool => string => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(option('a), Errors.t);

Validates the given Js.Json.t value at the given key, using the validation function.

A missing key or null value can be treated as None or returned as an error using the missingAsNone flag.

A null value can be treated as None or returned as an error using the nullAsNone flag.

Failed validation can be treated as None or returned as an error using the errorAsNone flag.

let validateArrayForKey: a. string => (int => json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(array('a), Errors.t);

Validates the given Js.Json.t value is an object with an array at the given key, then validates the array using the given validation function.

let validateListForKey: a. string => (int => json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(list('a), Errors.t);

Validates the given Js.Json.t value is an object with an array at the given key, then validates the array using the given validation function, returning the result in a list.

let validateObjectForKey: a. string => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t('a, Errors.t);

Validates the given Js.Json.t value is an object with an object at the given key, then validates the object using the given validation function.

module DSL: { ... };

Exposes a set of focused/abbreviated operations for use as a global or local open.