Module Relude_Js_Json.DSL

Exposes a set of focused/abbreviated operations for use as a global or local open.

The map/apply operators are provided for applicative style validation for decoding objects.

let (<$>): ('a => 'b) => ValidationE.Functor.t('a) => ValidationE.Functor.t('b);

let (<#>): ValidationE.Functor.t('a) => ('a => 'b) => ValidationE.Functor.t('b);

let (<*>): ValidationE.Apply.t(('a => 'b)) => ValidationE.Apply.t('a) => ValidationE.Apply.t('b);

let (>>=): ValidationE.Monad.t('a) => ('a => ValidationE.Monad.t('b)) => ValidationE.Monad.t('b);

module JE: { ... };

JSON encode utilities

module JD: { ... };

JSON decoding utilities