Module DSL.JD

* JSON decoding utilities

let null: json => Relude_Validation.t(unit, Errors.t);

* Validates the given Js.Json.t value is a null

let bool: json => Relude_Validation.t(bool, Errors.t);

* Validates the given Js.Json.t value is a bool

let int: json => Relude_Validation.t(int, Errors.t);

* Validates the given Js.Json.t value as a string

let float: json => Relude_Validation.t(float, Errors.t);

* Validates the given Js.Json.t value as a float

let string: json => Relude_Validation.t(string, Errors.t);

* Validates the given Js.Json.t value as a string

let opt: ?⁠errorAsNone:bool => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(option('a), Errors.t);

* Validates that the given Js.Json.t value is either null or can be validated using the given function. * * If the validation function fails, the error can either be returned as a successful None value, or the error * can be returned - depending on the errorAsNone flag.

let getAt: int => json => option(json);

* Gets the Js.Json.t value at the given index of a Js.Json.t array

let jsonAt: a. int => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t('a, Errors.t);

* Validates the Js.Json.t value at the given index with the given validation function

let nullAt: int => json => Relude_Validation.t(unit, Errors.t);

* Validates a null value at the given index of a Js.Json.t array value

let boolAt: int => json => Relude_Validation.t(bool, Errors.t);

* Validates a bool value at the given index of a Js.Json.t array value

let stringAt: int => json => Relude_Validation.t(string, Errors.t);

* Validates a string value at the given index of a Js.Json.t array value

let intAt: int => json => Relude_Validation.t(int, Errors.t);

* Validates an int value at the given index of a Js.Json.t array value

let floatAt: int => json => Relude_Validation.t(float, Errors.t);

* Validates a float value at the given index of a Js.Json.t array value

let optAt: ?⁠missingAsNone:bool => ?⁠nullAsNone:bool => ?⁠errorAsNone:bool => int => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(option('a), Errors.t);

* Validates that the Js.Json.t value at the given index is either null or can be validated using the given function. * * Bad indices can be treated as None or as an error using missingAsNone * Null value can be treated as None or as an error using nullAsNone * Failed validation can be treated as None or an error using errorAsNone

let arrayAt: a. int => (int => json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(array('a), Errors.t);

* Validates an array at the given index of a Js.Json.t array value

let objectAt: a. int => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t('a, Errors.t);

* Validates an obejct at the given index of a Js.Json.t array value

let array: a. (int => json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(array('a), Errors.t);

* Validates an Js.Json.t array using the given validation function

let list: a. (int => json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(list('a), Errors.t);

* Validates an Js.Json.t object using the given validation function

let getFor: string => json => option(json);

* Gets the Js.Json.t value for the given key of a Js.Json.t object value

let jsonFor: a. string => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t('a, Errors.t);

* Validates the Js.Json.t value for the given key, using the given validation function

let nullFor: string => json => Relude_Validation.t(unit, Errors.t);

* Validates a null for the given key of a Js.Json.t object value

let boolFor: string => json => Relude_Validation.t(bool, Errors.t);

* Validates a bool for the given key of a Js.Json.t object value

let stringFor: string => json => Relude_Validation.t(string, Errors.t);

* Validates a string for the given key of a Js.Json.t object value

let intFor: string => json => Relude_Validation.t(int, Errors.t);

* Validates an int for the given key of a Js.Json.t object value

let floatFor: string => json => Relude_Validation.t(float, Errors.t);

* Validates a float for the given key of a Js.Json.t object value

let optFor: ?⁠missingAsNone:bool => ?⁠nullAsNone:bool => ?⁠errorAsNone:bool => string => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(option('a), Errors.t);

* Validates that the Js.Json.t value at the given key is either null or can be validated using the given function. * * Bad indices can be treated as None or as an error using missingAsNone * Null value can be treated as None or as an error using nullAsNone * Failed validation can be treated as None or an error using errorAsNone

let arrayFor: a. string => (int => json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(array('a), Errors.t);

* Validates an array for the given key of a Js.Json.t object value

let listFor: a. string => (int => json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t(list('a), Errors.t);

* Validates an array for the given key of a Js.Json.t object value, as a list

let objectFor: string => (json => Relude_Validation.t('a, Errors.t)) => json => Relude_Validation.t('a, Errors.t);

* Validates an object for the given key of a Js.Json.t object value