Module DSL.JE

JSON encode utilities

let null: json;

* Creates a JSON null value

let bool: bool => json;

* Encodes a bool as a JSON bool value

let int: int => json;

* Encodes an int as a JSON number value

let float: float => json;

* Encodes a float as a JSON number value

let num: float => json;

* Encodes a float as a JSON number value * * Alias of `float`

let string: string => json;

* Encodes a string as a JSON string value

let opt: ('a => json) => option('a) => json;

* Encodes an option('a) as JSON, using the given encode function, or returning null for None.

let array: array(json) => json;

* Encodes an array(Js.Json.t) as a single Js.Json.t (array) value

let arrayBy: a. ('a => json) => array('a) => json;

* Maps a JSON-conversaion function over an array of values, and then encodes the result as a JSON array

let arrayOfDict: array(dict) => json;

* Encodes an array(Js.Dict.t(Js.Json.t)) into a Js.Json.t value

let arrayOfTuples: array((Js.Dict.key, json)) => json;

* Encodes an array of key/value pairs into a Js.Json.t value

let list: list(json) => json;

* Encodes a list of Js.Json.t values to a Js.Json.t value

let listBy: ('a => json) => list('a) => json;

* Maps a JSON-conversion function over a list of values, then encodes the result as a Js.Json.t array value

let listOfDict: list(dict) => json;

* Encodes a list Js.Dict.t(Js.Json.t) values into a Js.Json.t array value

let listOfTuples: list((Js.Dict.key, json)) => json;

* Encodes a list of key/value pairs into a Js.Json.t array value

let dict: dict => json;

* Encodes a dict of Js.Json.t values into a Js.Json.t value