Module Relude_List_Base

let cons: a. 'a => list('a) => list('a);

List.cons prepends the given item to the start of the given list. This preserves the memory used by the existing list.

Running time: O(1)

List.cons("x", ["y", "z"]) == ["x", "y", "z"];
List.cons(0, []) == [0];

let prepend: a. 'a => list('a) => list('a);

List.prepend is an alias for cons.

let uncons: a. list('a) => option(('a, list('a)));

List.uncons splits a list into head and tail parts. This returns None if the list is empty.

Running time: O(1)

List.uncons(["a", "b", "c"]) == Some(("a", ["b", "c"]));
List.uncons([0]) == Some((0, []));
List.uncons([]) == None;

let append: a. 'a => list('a) => list('a);

List.append adds an item to the end of the list. This requires a full copy to be made of the provided list, which means that the returned list will have no relationship in memory to the original list.

Running time: O(n)

List.append(4, [1, 2, 3]) == [1, 2, 3, 4];

let repeat: a. int => 'a => list('a);

List.repeat accepts a count and a value, and it creates a new list that contains the provided value, duplicated count times.

Note that a negative count will return an empty list.

Running time: O(n) (where n is the provided count)

List.repeat(3, "hello") == ["hello", "hello", "hello"];
List.repeat(-1, 42) == [];

let makeWithIndex: a. int => (int => 'a) => list('a);

Makes a list by mapping a function over a range of ints from 0 to [n - 1].

let mapWithIndex: a b. ('a => int => 'b) => list('a) => list('b);

Maps an indexed function over the values of a list to produce a new list.

let reverse: a. list('a) => list('a);

Reverses the given list

let shuffle: a. list('a) => list('a);

Shuffles the given list to create a new list

let isEmpty: a. list('a) => bool;

Indicates if the given list is empty (length == 0).

let isNotEmpty: a. list('a) => bool;

Indicates if the given list is non-empty (length > 0).

let at: a. int => list('a) => option('a);

Gets the value at the given index of the list, or None if the index is out of range.

let head: a. list('a) => option('a);

Gets the first item of the list, or None if the list is empty.

let tail: a. list('a) => option(list('a));

Gets all but the first items in the list, or None if the list is empty.

let tailOrEmpty: a. list('a) => list('a);

Gets all but the first items in the list, or [] if the list is empty.

let init: a. list('a) => option(list('a));

Gets all but the last item in the list, or None if the list is empty

let initOrEmpty: list('a) => list('a);

Gets all but the last item in the list, or [] if the list is empty

let last: a. list('a) => option('a);

Gets the last item of the list, or None if the list is empty.

let take: a. int => list('a) => list('a);

Creates a new list by taking up to n items from the given list.

let takeExactly: a. int => list('a) => option(list('a));

Creates a new list by taking exactly n items from the given list. If there are not at least n items, the result is None.

let takeWhile: a. ('a => bool) => list('a) => list('a);

Creates a new list by taking items from the list until an item is reached that does not satisfy the given predicate.

let drop: a. int => list('a) => list('a);

Creates a new list by dropping up to n items from the given list.

let dropExactly: a. int => list('a) => option(list('a));

Creates a new list by dropping exactly n items from the given list. If there are fewer than n items, None is returned.

let dropWhile: a. ('a => bool) => list('a) => list('a);

Creates a new list by dropping items from the list until an item is reached which does not satisfy the given predicate.

let filter: a. ('a => bool) => list('a) => list('a);

Creates a new list containing only the items from the given list that satisfy the given predicate.

let keep: a. ('a => bool) => list('a) => list('a);

Alias of filter

let filterWithIndex: a. ('a => int => bool) => list('a) => list('a);

Creates a new list containing only the items from the given list that satisfy the given indexed predicate.

let keepWithIndex: a. ('a => int => bool) => list('a) => list('a);

Alias of filterWithIndex

let filterNot: a. ('a => bool) => list('a) => list('a);

Creates a new list containing only the items from the given list that do not satisfy the given predicate.

let reject: a. ('a => bool) => list('a) => list('a);

Alias of filterNot

let filterNotWithIndex: a. ('a => int => bool) => list('a) => list('a);

Creates a new list containing only the items from the given list that do not satisfy the given indexed predicate.

let rejectWithIndex: a. ('a => int => bool) => list('a) => list('a);

Alias of filterNotWithIndex

let mapOption: a b. ('a => option('b)) => list('a) => list('b);

Maps a option-creating function over the list, and keeps the Some values without the Some structure.

let catOptions: a. list(option('a)) => list('a);

Keeps all the Some values from the list, and removes the Some structure.

let partition: a. ('a => bool) => list('a) => (list('a), list('a));

Partitions the given list into two lists, one (left-side) containing values that satisfy the given predicate and one (right-side) containing values that do not satisfy the predicate.

let splitAt: a. int => list('a) => option((list('a), list('a)));

Splits a list into two lists at the given index. None is returned if the index is out of range.

let prependToAll: a. 'a => list('a) => list('a);

Creates a new list by prepending the given value to each item in the list.

let intersperse: a. 'a => list('a) => list('a);

Creates a new list by inserting the given item in between each item in the list.

let replicate: a. int => list('a) => list('a);

Creates a new list by copying the given list n times.

let zip: a b. list('a) => list('b) => list(('a, 'b));

Combines two lists pair-wise into a list of tuple-2

let zipWith: a b c. ('a => 'b => 'c) => list('a) => list('b) => list('c);

Combines two lists using a function to combine pair-wise values.

let zipWithIndex: a. list('a) => list(('a, int));

Creates a new list with each item paired with its index in the list.

let unzip: a b. list(('a, 'b)) => (list('a), list('b));

Creates two lists by splitting a list of tuple-2 on the left and right.

let sortWithInt: a. ('a => 'a => int) => list('a) => list('a);

Sorts a list with the given int-based compare function.

let sortBy: a. ('a => 'a => BsBastet.Interface.ordering) => list('a) => list('a);

Sorts a list with the given ordering-based compare function.

let sort: (module BsBastet.Interface.ORD with type t = 'a) => list('a) => list('a);

Sorts a list with the given ORD module.

let distinctBy: a. ('a => 'a => bool) => list('a) => list('a);

Creates a new list containing only the distinct items of the given list, using the given equality function.

let removeFirstBy: a. ('a => 'a => bool) => 'a => list('a) => list('a);

Removes the first occurrence of the given item from the list, based on the given equality function.

let removeEachBy: a. ('a => 'a => bool) => 'a => list('a) => list('a);

Removes all occurrences of the given item from the list, based on the given equality function.

let distinct: (module BsBastet.Interface.EQ with type t = 'a) => list('a) => list('a);

Creates a new list containing only the distinct values of the list, based on the given equality function.

let removeFirst: (module BsBastet.Interface.EQ with type t = 'a) => 'a => list('a) => list('a);

Removes the first occurrence of the given item from the list, based on the given EQ module.

let removeEach: (module BsBastet.Interface.EQ with type t = 'a) => 'a => list('a) => list('a);

Removes all occurrences of the given item from the list, based on the given EQ module.

let replaceAt: a. int => 'a => list('a) => list('a);

Creates a new list that replaces the item at the given index with the given value. If the index is out of range, no replacement is made.

let scanLeft: ('b => 'a => 'b) => 'b => list('a) => list('b);

Similar to foldLeft, but collects the results from each iteration, rather than accumulating a single final result.

let scanRight: ('a => 'b => 'b) => 'b => list('a) => list('b);

Similar to foldRight, but collects the results from each iteration, rather than accumulating a single final result.

let insertAt: a. int => 'a => list('a) => list('a);

Creates a new list that inserts the given value at the given index. If the index is out of range, no insertion is made.

let updateAt: a. int => ('a => 'a) => list('a) => list('a);

Creates a new list that modifies the value at the given index with the given function. If the index is out of range, no change is made.

let swapAt: a. int => int => list('a) => list('a);

Creates a new list with the elements at the two given indexes swapped. If either index is out of range, no change is made.

let removeAt: a. int => list('a) => list('a);

Creates a new list without the element at the given index. If the index is out of range, no change is made.

let chunk: a. int => list('a) => list(list('a));

Creates a list of elements split into groups the length of size. If the list can't be split evenly, the final chunk will be the remaining elements.