Module Relude_List_Instances

let concat: a. list('a) => list('a) => list('a);

Concatenates two lists with the left-side first and the right-side last.

module SemigroupAny: BsBastet.Interface.SEMIGROUP_ANY with type SemigroupAny.t('a) = list('a);

include { ... };

let concatNamed: prefix:SemigroupAny.t('a) => SemigroupAny.t('a) => SemigroupAny.t('a);

let empty: a. list('a);

Creates an empty list.

module MonoidAny: BsBastet.Interface.MONOID_ANY with type MonoidAny.t('a) = list('a);

include { ... };

let guard: bool => MonoidAny.t('a) => MonoidAny.t('a);

let power: MonoidAny.t('a) => int => MonoidAny.t('a);

let map: ('a => 'b) => BsBastet.List.Functor.t('a) => BsBastet.List.Functor.t('b);

Maps a pure function over a list

module Functor: BsBastet.Interface.FUNCTOR with type Functor.t('a) = list('a);

include { ... };

module BsFunctorExtensions: { ... };

let flipMap: Functor.t('a) => ('a => 'b) => Functor.t('b);

let void: Functor.t('a) => Functor.t(unit);

let voidRight: 'a => Functor.t('b) => Functor.t('a);

let voidLeft: Functor.t('a) => 'b => Functor.t('b);

let flap: Functor.t(('a => 'b)) => 'a => Functor.t('b);

let apply: BsBastet.List.Apply.t(('a => 'b)) => BsBastet.List.Apply.t('a) => BsBastet.List.Apply.t('b);

Applies a list of functions to a list of values.

module Apply: BsBastet.Interface.APPLY with type Apply.t('a) = list('a);

include { ... };

module BsApplyExtensions: { ... };

let applyFirst: Apply.t('a) => Apply.t('b) => Apply.t('a);

let applySecond: Apply.t('a) => Apply.t('b) => Apply.t('b);

let map2: ('a => 'b => 'c) => Apply.t('a) => Apply.t('b) => Apply.t('c);

let map3: ('a => 'b => 'c => 'd) => Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d);

let map4: ('a => 'b => 'c => 'd => 'e) => Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d) => Apply.t('e);

let map5: ('a => 'b => 'c => 'd => 'e => 'f) => Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d) => Apply.t('e) => Apply.t('f);

let tuple2: Apply.t('a) => Apply.t('b) => Apply.t(('a, 'b));

let tuple3: Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t(('a, 'b, 'c));

let tuple4: Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d) => Apply.t(('a, 'b, 'c, 'd));

let tuple5: Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d) => Apply.t('e) => Apply.t(('a, 'b, 'c, 'd, 'e));

let mapTuple2: ('a => 'b => 'c) => (Apply.t('a), Apply.t('b)) => Apply.t('c);

let mapTuple3: ('a => 'b => 'c => 'd) => (Apply.t('a), Apply.t('b), Apply.t('c)) => Apply.t('d);

let mapTuple4: ('a => 'b => 'c => 'd => 'e) => (Apply.t('a), Apply.t('b), Apply.t('c), Apply.t('d)) => Apply.t('e);

let mapTuple5: ('a => 'b => 'c => 'd => 'e => 'f) => (Apply.t('a), Apply.t('b), Apply.t('c), Apply.t('d), Apply.t('e)) => Apply.t('f);

let pure: 'a => BsBastet.List.Applicative.t('a);

Lifts a single pure value into a list of one value.

module Applicative: BsBastet.Interface.APPLICATIVE with type Applicative.t('a) = list('a);

include { ... };

module BsApplicativeExtensions: { ... };

let liftA1: ('a => 'b) => Applicative.t('a) => Applicative.t('b);

let when_: bool => Applicative.t(unit) => Applicative.t(unit);

let unless: bool => Applicative.t(unit) => Applicative.t(unit);

let all: list(Applicative.t('a)) => Applicative.t(list('a));

let bind: BsBastet.List.Monad.t('a) => ('a => BsBastet.List.Monad.t('b)) => BsBastet.List.Monad.t('b);

Maps a monadic function over a list of values and flattens the result.

module Monad: BsBastet.Interface.MONAD with type Monad.t('a) = list('a);

include { ... };

module BsMonadExtensions: { ... };

let flatMap: ('a => Monad.t('b)) => Monad.t('a) => Monad.t('b);

let flatten: Monad.t(Monad.t('a)) => Monad.t('a);

let composeKleisli: ('a => Monad.t('b)) => ('b => Monad.t('c)) => 'a => Monad.t('c);

let flipComposeKleisli: ('b => Monad.t('c)) => ('a => Monad.t('b)) => 'a => Monad.t('c);

let liftM1: ('a => 'b) => Monad.t('a) => Monad.t('b);

let when_: Monad.t(bool) => Monad.t(unit) => Monad.t(unit);

let unless: Monad.t(bool) => Monad.t(unit) => Monad.t(unit);

let alt: BsBastet.List.Alt.t('a) => BsBastet.List.Alt.t('a) => BsBastet.List.Alt.t('a);

alt for lists concatenates the lists.

module Alt: BsBastet.Interface.ALT with type Alt.t('a) = list('a);

include { ... };

let orElse: fallback:Alt.t('a) => Alt.t('a) => Alt.t('a);

module Plus: BsBastet.Interface.PLUS with type Plus.t('a) = list('a);

include { ... };

module Alternative: BsBastet.Interface.ALTERNATIVE with type Alternative.t('a) = list('a);

include { ... };

let foldLeft: ('a => 'b => 'a) => 'a => BsBastet.List.Foldable.t('b) => 'a;

Folds a list from left-to-right into a single value using an accumulator

let foldRight: ('a => 'b => 'b) => 'b => BsBastet.List.Foldable.t('a) => 'b;

Folds a list from right-to-left into a single value using an accumulator

module Foldable: BsBastet.Interface.FOLDABLE with type Foldable.t('a) = list('a);

include { ... };

module BsFoldableExtensions: { ... };

let any: ('a => bool) => Foldable.t('a) => bool;

let all: ('a => bool) => Foldable.t('a) => bool;

let containsBy: ('a => 'a => bool) => 'a => Foldable.t('a) => bool;

let contains: (module BsBastet.Interface.EQ with type t = 'a) => 'a => Foldable.t('a) => bool;

let indexOfBy: ('a => 'a => bool) => 'a => Foldable.t('a) => option(int);

let indexOf: (module BsBastet.Interface.EQ with type t = 'a) => 'a => Foldable.t('a) => option(int);

let minBy: ('a => 'a => BsBastet.Interface.ordering) => Foldable.t('a) => option('a);

let min: (module BsBastet.Interface.ORD with type t = 'a) => Foldable.t('a) => option('a);

let maxBy: ('a => 'a => BsBastet.Interface.ordering) => Foldable.t('a) => option('a);

let max: (module BsBastet.Interface.ORD with type t = 'a) => Foldable.t('a) => option('a);

let countBy: ('a => bool) => Foldable.t('a) => int;

let length: Foldable.t('a) => int;

let size: Foldable.t('a) => int;

let count: Foldable.t('a) => int;

let forEach: ('a => unit) => Foldable.t('a) => unit;

let forEachWithIndex: ('a => int => unit) => Foldable.t('a) => unit;

let find: ('a => bool) => Foldable.t('a) => option('a);

let findWithIndex: ('a => int => bool) => Foldable.t('a) => option('a);

let toList: Foldable.t('a) => list('a);

let toArray: Foldable.t('a) => array('a);

module FoldableSemigroupExtensions: (S: BsBastet.Interface.SEMIGROUP) => { ... };

module FoldableMonoidExtensions: (M: BsBastet.Interface.MONOID) => { ... };

let foldMap: (module BsBastet.Interface.MONOID with type t = 'a) => ('b => 'a) => Foldable.t('b) => 'a;

let foldWithMonoid: (module BsBastet.Interface.MONOID with type t = 'a) => Foldable.t('a) => 'a;

let intercalate: (module BsBastet.Interface.MONOID with type t = 'a) => 'a => Foldable.t('a) => 'a;

module FoldableApplicativeExtensions: (A: BsBastet.Interface.APPLICATIVE) => { ... };

module FoldableMonadExtensions: (M: BsBastet.Interface.MONAD) => { ... };

module FoldableEqExtensions: (E: BsBastet.Interface.EQ) => { ... };

module FoldableOrdExtensions: (O: BsBastet.Interface.ORD) => { ... };

let unfold: ('a => option(('a, 'a))) => 'a => BsBastet.List.Unfoldable.t('a);

Create a list by recursively unfolding with a generator function and seed value

module Unfoldable: BsBastet.Interface.UNFOLDABLE with type Unfoldable.t('a) = list('a);

include { ... };

module Traversable: BsBastet.List.TRAVERSABLE_F;

let eqBy: ('a => 'a => bool) => list('a) => list('a) => bool;

Compares two lists for length and pair-wise equality using the given equality function

let eq: (module BsBastet.Interface.EQ with type t = 'a) => list('a) => list('a) => bool;

Compares two lists for length and pair-wise equality using the given EQ module

module Eq: (EqA: BsBastet.Interface.EQ) => { ... };

let showBy: ('a => string) => list('a) => string;

Converts a list to a string using the given show function

let show: (module BsBastet.Interface.SHOW with type t = 'a) => list('a) => string;

Converts a list to a string using the given SHOW module

module Show: (ShowA: BsBastet.Interface.SHOW) => { ... };

let fromArray: array('a) => Belt.List.t('a);

Converts the given array to a list

let toArray: Belt.List.t('a) => array('a);

Converts the given list to an array

module IsoArray: Relude_Interface.ISO_ARRAY with type IsoArray.t('a) = list('a);