Module Relude_List_Specializations.Float

List extensions for list(float)

include { ... };

let contains: Relude_Float.Ord.t => Relude_List_Instances.Foldable.t(Relude_Float.Ord.t) => bool;

let indexOf: Relude_Float.Ord.t => Relude_List_Instances.Foldable.t(Relude_Float.Ord.t) => option(int);

let distinct: list(Relude_Float.Ord.t) => list(Relude_Float.Ord.t);

let removeFirst: Relude_Float.Ord.t => list(Relude_Float.Ord.t) => list(Relude_Float.Ord.t);

let removeEach: Relude_Float.Ord.t => list(Relude_Float.Ord.t) => list(Relude_Float.Ord.t);

let eq: list(Relude_Float.Ord.t) => list(Relude_Float.Ord.t) => bool;

let min: Relude_List_Instances.Foldable.t(Relude_Float.Ord.t) => option(Relude_Float.Ord.t);

let max: Relude_List_Instances.Foldable.t(Relude_Float.Ord.t) => option(Relude_Float.Ord.t);

let sort: list(Relude_Float.Ord.t) => list(Relude_Float.Ord.t);

let sum: list(float) => float;

Finds the sum of all the floats in the list

let product: list(float) => float;

Finds the product of all the floats in the list