Module Relude_List_Specializations.Int

List extensions for list(int)

include { ... };

let contains: Relude_Int.Ord.t => Relude_List_Instances.Foldable.t(Relude_Int.Ord.t) => bool;

let indexOf: Relude_Int.Ord.t => Relude_List_Instances.Foldable.t(Relude_Int.Ord.t) => option(int);

let distinct: list(Relude_Int.Ord.t) => list(Relude_Int.Ord.t);

let removeFirst: Relude_Int.Ord.t => list(Relude_Int.Ord.t) => list(Relude_Int.Ord.t);

let removeEach: Relude_Int.Ord.t => list(Relude_Int.Ord.t) => list(Relude_Int.Ord.t);

let eq: list(Relude_Int.Ord.t) => list(Relude_Int.Ord.t) => bool;

let min: Relude_List_Instances.Foldable.t(Relude_Int.Ord.t) => option(Relude_Int.Ord.t);

let max: Relude_List_Instances.Foldable.t(Relude_Int.Ord.t) => option(Relude_Int.Ord.t);

let sort: list(Relude_Int.Ord.t) => list(Relude_Int.Ord.t);

let sum: list(int) => int;

Finds the sum of all the ints in the list

let product: list(int) => int;

Finds the product of all the ints in the list