Module Relude_List_Specializations.ListEqExtensions

* List extensions for when you have an EQ instance.

Parameters

E: BsBastet.Interface.EQ

Signature

include { ... };

let contains: E.t => Relude_List_Instances.Foldable.t(E.t) => bool;

let indexOf: E.t => Relude_List_Instances.Foldable.t(E.t) => option(int);

let distinct: list(E.t) => list(E.t);

* Gets the distinct items of the list, based on the given EQ module

let removeFirst: E.t => list(E.t) => list(E.t);

* Removes the first item of the list which equals the given item, based on the given EQ module

let removeEach: E.t => list(E.t) => list(E.t);

* Removes all items of the list which equal the given item, based on the given EQ module

let eq: list(E.t) => list(E.t) => bool;

* Indicates if all pairwise items in the given lists are equal, using the given EQ module