Module Relude_List_Specializations.String

include { ... };

let contains: Relude_String.Ord.t => Relude_List_Instances.Foldable.t(Relude_String.Ord.t) => bool;

let indexOf: Relude_String.Ord.t => Relude_List_Instances.Foldable.t(Relude_String.Ord.t) => option(int);

let distinct: list(Relude_String.Ord.t) => list(Relude_String.Ord.t);

let removeFirst: Relude_String.Ord.t => list(Relude_String.Ord.t) => list(Relude_String.Ord.t);

let removeEach: Relude_String.Ord.t => list(Relude_String.Ord.t) => list(Relude_String.Ord.t);

let eq: list(Relude_String.Ord.t) => list(Relude_String.Ord.t) => bool;

let min: Relude_List_Instances.Foldable.t(Relude_String.Ord.t) => option(Relude_String.Ord.t);

let max: Relude_List_Instances.Foldable.t(Relude_String.Ord.t) => option(Relude_String.Ord.t);

let sort: list(Relude_String.Ord.t) => list(Relude_String.Ord.t);

include { ... };

module BsFoldableMonoidExtensions: { ... };

let foldMap: ('a => Relude_String.Monoid.t) => Relude_List_Instances.Foldable.t('a) => Relude_String.Monoid.t;

let foldWithMonoid: Relude_List_Instances.Foldable.t(Relude_String.Monoid.t) => Relude_String.Monoid.t;

let intercalate: Relude_String.Monoid.t => Relude_List_Instances.Foldable.t(Relude_String.Monoid.t) => Relude_String.Monoid.t;

let join: list(string) => string;

* Joins a list of strings using the empty string "" as a delimiter

let joinWith: string => list(string) => string;

* Joins a list of strings using the given delimiter

let distinct: list(string) => list(string);

* Remove all duplicate entries from a list of strings. * * Instead of passing String.Eq to the underlying List.distinct function, we * have a special-case implementation here that uses Js.Dict for faster * comparison. This is O(n) instead of O(n^2).