Module Relude_List_Specializations.Validation

* List extensions for `Validation.t('a, 'e)`

module Traversable: (Errors: BsBastet.Interface.SEMIGROUP_ANY) => (Error: BsBastet.Interface.TYPE) => { ... };

module TraversableWithErrorsAsList: (Error: BsBastet.Interface.TYPE) => { ... };

module TraversableWithErrorsAsListOfStrings: { ... };

module TraversableWithErrorsAsNonEmptyList: (Error: BsBastet.Interface.TYPE) => { ... };

let traverse: ('a => Pervasives.result('b, 'c)) => list('a) => Relude_Validation.t(list('b), Relude_NonEmpty.List.t('c));

* This is a streamlined definition of traverse which allows you to return a * result for each item in the list, and all errors are collected in a * `NonEmptyList` of your error type, using applicative semantics for * Validation.