Module Relude_NonEmpty

module WithSequence: (TailSequence: Relude_Interface.SEQUENCE) => { ... };

Creates a NonEmpty module with the given SEQUENCE module to handle the tail part.

module List: { ... };

A NonEmpty implemented using a list as the tail sequence

module Array: { ... };

A NonEmpty implemented using an array as the tail sequence