Module Relude_Option_Instances

let compose: a b c. option(('b => 'c)) => option(('a => 'b)) => option(('a => 'c));

module Semigroupoid: BsBastet.Interface.SEMIGROUPOID with type Semigroupoid.t('a, 'b) = option(('a => 'b));

include { ... };

let andThen: Semigroupoid.t('a, 'b) => Semigroupoid.t('b, 'c) => Semigroupoid.t('a, 'c);

let map: a b. ('a => 'b) => option('a) => option('b);

`map(f, opt)`, when `opt` is `Some(v)`, returns `Some(f(v))`. When `opt` is `None`, it returns `None`.

map(x => x * x, Some(12)) == Some(144);
map(x => x * x, None) == None;

module Functor: BsBastet.Interface.FUNCTOR with type Functor.t('a) = option('a);

include { ... };

module BsFunctorExtensions: { ... };

let flipMap: Functor.t('a) => ('a => 'b) => Functor.t('b);

let void: Functor.t('a) => Functor.t(unit);

let voidRight: 'a => Functor.t('b) => Functor.t('a);

let voidLeft: Functor.t('a) => 'b => Functor.t('b);

let flap: Functor.t(('a => 'b)) => 'a => Functor.t('b);

let apply: a b. option(('a => 'b)) => option('a) => option('b);

`apply(optFcn, optVal)` returns `Some(f(v))` if `optFcn` is `Some(f)` and `optVal` is `Some(v)`. In all other cases, `apply()` returns `None`.

let square = (x) => {x * x};
apply(Some(square), Some(12)) == Some(144);
apply(Some(square), None) == None;
apply(None, Some(12)) == None;
apply(None, None) == None;

module Apply: BsBastet.Interface.APPLY with type Apply.t('a) = option('a);

include { ... };

module BsApplyExtensions: { ... };

let applyFirst: Apply.t('a) => Apply.t('b) => Apply.t('a);

let applySecond: Apply.t('a) => Apply.t('b) => Apply.t('b);

let map2: ('a => 'b => 'c) => Apply.t('a) => Apply.t('b) => Apply.t('c);

let map3: ('a => 'b => 'c => 'd) => Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d);

let map4: ('a => 'b => 'c => 'd => 'e) => Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d) => Apply.t('e);

let map5: ('a => 'b => 'c => 'd => 'e => 'f) => Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d) => Apply.t('e) => Apply.t('f);

let tuple2: Apply.t('a) => Apply.t('b) => Apply.t(('a, 'b));

let tuple3: Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t(('a, 'b, 'c));

let tuple4: Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d) => Apply.t(('a, 'b, 'c, 'd));

let tuple5: Apply.t('a) => Apply.t('b) => Apply.t('c) => Apply.t('d) => Apply.t('e) => Apply.t(('a, 'b, 'c, 'd, 'e));

let mapTuple2: ('a => 'b => 'c) => (Apply.t('a), Apply.t('b)) => Apply.t('c);

let mapTuple3: ('a => 'b => 'c => 'd) => (Apply.t('a), Apply.t('b), Apply.t('c)) => Apply.t('d);

let mapTuple4: ('a => 'b => 'c => 'd => 'e) => (Apply.t('a), Apply.t('b), Apply.t('c), Apply.t('d)) => Apply.t('e);

let mapTuple5: ('a => 'b => 'c => 'd => 'e => 'f) => (Apply.t('a), Apply.t('b), Apply.t('c), Apply.t('d), Apply.t('e)) => Apply.t('f);

let pure: a. 'a => option('a);

`pure(v)` returns `Some(v)`.

pure("Reason") == Some("Reason");

module Applicative: BsBastet.Interface.APPLICATIVE with type Applicative.t('a) = option('a);

include { ... };

module BsApplicativeExtensions: { ... };

let liftA1: ('a => 'b) => Applicative.t('a) => Applicative.t('b);

let when_: bool => Applicative.t(unit) => Applicative.t(unit);

let unless: bool => Applicative.t(unit) => Applicative.t(unit);

let all: list(Applicative.t('a)) => Applicative.t(list('a));

let bind: a b. option('a) => ('a => option('b)) => option('b);

`bind(opt, f)` returns `f(v)` if `opt` is `Some(v)`, `None` otherwise. In this case, `f` is a function that takes a non-`option` argument and returns an `option` result.

let reciprocalOpt = (x) => { x == 0.0 ? None : Some(1.0 /. x) };
bind(Some(2.0), reciprocalOpt) == Some(0.5);
bind(Some(0.0), reciprocalOpt) == None;
bind(None, reciprocalOpt) == None;

`bind` is the same as `flatMap`, but with the arguments in the reverse order.

module Monad: BsBastet.Interface.MONAD with type Monad.t('a) = option('a);

include { ... };

module BsMonadExtensions: { ... };

let flatMap: ('a => Monad.t('b)) => Monad.t('a) => Monad.t('b);

let flatten: Monad.t(Monad.t('a)) => Monad.t('a);

let composeKleisli: ('a => Monad.t('b)) => ('b => Monad.t('c)) => 'a => Monad.t('c);

let flipComposeKleisli: ('b => Monad.t('c)) => ('a => Monad.t('b)) => 'a => Monad.t('c);

let liftM1: ('a => 'b) => Monad.t('a) => Monad.t('b);

let when_: Monad.t(bool) => Monad.t(unit) => Monad.t(unit);

let unless: Monad.t(bool) => Monad.t(unit) => Monad.t(unit);

let align: a b. option('a) => option('b) => option(Relude_Ior_Type.t('a, 'b));

let alignWith: a b c. (Relude_Ior_Type.t('a, 'b) => 'c) => option('a) => option('b) => option('c);

module Semialign: Relude_Interface.SEMIALIGN with type Semialign.t('a) = option('a);

include { ... };

module Align: Relude_Interface.ALIGN with type Align.t('a) = option('a);

include { ... };

let foldLeft: a b. ('b => 'a => 'b) => 'b => option('a) => 'b;

`foldLeft(f, init, opt)` takes as its first argument a function `f` that has two arguments. The first argument is an “accumulator“ of the same type as `init` (the initial value of the accumulator), and the second is of the same type as the value wrapped in `opt`. `f()` returns a value of the same type as `init`.

If `opt` is `Some(v)`, `foldLeft()` returns `f(accumulator,v)`. If `opt` is `None`, `foldLeft()` returns `accumulator`.

let addLength = (accum, str) => {accum + Js.String.length(str)};
foldLeft(addLength, 0, Some("Reason")) == 6;
foldLeft(addLength, 0, None) == 0;

let foldRight: a b. ('a => 'b => 'b) => 'b => option('a) => 'b;

`foldRight(f, init, opt)` takes as its first argument a function `f` that has two arguments. The first argument is of the same type as the value wrapped in `opt`, and the second is an “accumulator“ of the same type as `init` (the initial value of the accumulator). `f()` returns a value of the same type as `init`.

If `opt` is `Some(v)`, `foldLeft()` returns `f(v, accumulator)`. If `opt` is `None`, `foldLeft()` returns `accumulator`.

let addLength = (str, accum) => {accum + Js.String.length(str)};
foldRight(addLength, 0, Some("Reason")) == 6;
foldRight(addLength, 0, None) == 0;

module Foldable = BsBastet.Option.Foldable;

include { ... };

module BsFoldableExtensions: { ... };

let any: ('a => bool) => Foldable.t('a) => bool;

let all: ('a => bool) => Foldable.t('a) => bool;

let containsBy: ('a => 'a => bool) => 'a => Foldable.t('a) => bool;

let contains: (module BsBastet.Interface.EQ with type t = 'a) => 'a => Foldable.t('a) => bool;

let indexOfBy: ('a => 'a => bool) => 'a => Foldable.t('a) => option(int);

let indexOf: (module BsBastet.Interface.EQ with type t = 'a) => 'a => Foldable.t('a) => option(int);

let minBy: ('a => 'a => BsBastet.Interface.ordering) => Foldable.t('a) => option('a);

let min: (module BsBastet.Interface.ORD with type t = 'a) => Foldable.t('a) => option('a);

let maxBy: ('a => 'a => BsBastet.Interface.ordering) => Foldable.t('a) => option('a);

let max: (module BsBastet.Interface.ORD with type t = 'a) => Foldable.t('a) => option('a);

let countBy: ('a => bool) => Foldable.t('a) => int;

let length: Foldable.t('a) => int;

let size: Foldable.t('a) => int;

let count: Foldable.t('a) => int;

let forEach: ('a => unit) => Foldable.t('a) => unit;

let forEachWithIndex: ('a => int => unit) => Foldable.t('a) => unit;

let find: ('a => bool) => Foldable.t('a) => option('a);

let findWithIndex: ('a => int => bool) => Foldable.t('a) => option('a);

let toList: Foldable.t('a) => list('a);

let toArray: Foldable.t('a) => array('a);

module FoldableSemigroupExtensions: (S: BsBastet.Interface.SEMIGROUP) => { ... };

module FoldableMonoidExtensions: (M: BsBastet.Interface.MONOID) => { ... };

let foldMap: (module BsBastet.Interface.MONOID with type t = 'a) => ('b => 'a) => Foldable.t('b) => 'a;

let foldWithMonoid: (module BsBastet.Interface.MONOID with type t = 'a) => Foldable.t('a) => 'a;

let intercalate: (module BsBastet.Interface.MONOID with type t = 'a) => 'a => Foldable.t('a) => 'a;

module FoldableApplicativeExtensions: (A: BsBastet.Interface.APPLICATIVE) => { ... };

module FoldableMonadExtensions: (M: BsBastet.Interface.MONAD) => { ... };

module FoldableEqExtensions: (E: BsBastet.Interface.EQ) => { ... };

module FoldableOrdExtensions: (O: BsBastet.Interface.ORD) => { ... };

let alt: a. option('a) => option('a) => option('a);

`alt(opt1, opt2)` returns `opt1` if it is of the form `Some(v)`; otherwise it returns `opt2`.

Note: the value to check is on the left, and the fallback value is on the right.

alt(Some(3), Some(4)) == Some(3);
alt(Some(3), None) == Some(3);
alt(None, Some(4)) == Some(4);
alt(None, None) == None;

let altLazy: a. option('a) => (unit => option('a)) => option('a);

Lazy version of `alt()` (doesn't evaluate the second argument unless needed

Note: the value to check is on the left, and the fallback value function is on the right.

module Semigroup_Any: BsBastet.Interface.SEMIGROUP_ANY;

module Monoid_Any: { ... };

module Alt = BsBastet.Option.Alt;

include { ... };

let orElse: fallback:Alt.t('a) => Alt.t('a) => Alt.t('a);

module Plus = BsBastet.Option.Plus;

include { ... };

module Alternative = BsBastet.Option.Alternative;

include { ... };

module Traversable: BsBastet.Option.TRAVERSABLE_F;

let eqBy: ('a => 'a => bool) => option('a) => option('a) => bool;

In `eqBy(f, opt1, opt2)`, `f` is a function that compares two arguments for equality and returns `true` if they are equal. If `opt1` and `opt2` are `Some(v1)` and `Some(v2)`, then `eqBy()` returns `f(v1, v2)`. If both `opt1` and `opt2` are `None`, then `eqBy()` also returns `true`. In all other cases, the result is `false`.

let clockEqual = (a, b) => {a mod 12 == b mod 12};
eqBy(clockEqual, Some(3), Some(15)) == true;
eqBy(clockEqual, Some(3), Some(16)) == false;
eqBy(clockEqual, None, Some(15)) == false;
eqBy(clockEqual, Some(3), None) == false;
eqBy(clockEqual, None, None) == true;

let eq: (module BsBastet.Interface.EQ with type t = 'a) => option('a) => option('a) => bool;

`eq((module M), opt1, opt2)`compares its two arguments for equality and returns `true` if they are equal. If `opt1` and `opt2` are `Some(v1)` and `Some(v2)`, then `eq()` returns the result of comparing two values of their type for equality according to `module M`. If both `opt1` and `opt2` are `None`, then `eq()` also returns `true`. In all other cases, the result is `false`.

`module M` must implement:

a type `t`

a function `eq(t, t)` that evaluates equality and returns a boolean

module ClockEq = {
  type t = int;
  let eq = (a, b) => {a mod 12 == b mod 12};
};
eq((module ClockEq), Some(3), Some(15)) == true;
eq((module ClockEq), Some(3), Some(16)) == false;
eq((module ClockEq), None, Some(16)) == false;
eq((module ClockEq), Some(3), None) == false;
eq((module ClockEq), None, None) == true;

module Eq: BsBastet.Option.EQ_F;

module Ord: BsBastet.Option.ORD_F;

let showBy: a. ('a => string) => option('a) => string;

Converts the option to a string using the given show function

let show: (module BsBastet.Interface.SHOW with type t = 'a) => option('a) => string;

Converts the option to a string using the given SHOW module

module Show = BsBastet.Option.Show;

module WithSemigroup: (S: BsBastet.Interface.SEMIGROUP) => { ... };

module WithApplicative: (A: BsBastet.Interface.APPLICATIVE) => { ... };