Module Relude_Ord

type nonrec ordering = BsBastet.Interface.ordering;

* Contains functions and modules for to help with comparing values

* The ordering type represents the result of a comparison - `less_than, `equal_to, or `greater_than

type compare('a) = 'a => 'a => ordering;

* The compare function is the heart of all comparison functions

let by: a b. ('b => 'a) => compare('a) => compare('b);

* Converts an compare function for a type 'a to a compare function of type 'b * using a function from 'b => 'a. * * If we know how to compare 'as and we know how to convert 'b to 'a, we can * compare 'bs by first converting our 'bs to 'as, then comparing them with the 'a * compare function. * * Example: * ``` * let userCompare: compare(user) = Relude.Ord.by(user => user.name, Relude.String.compare); * ``` * * This is the contravariant map for the compare function.

let cmap: ('a => 'b) => compare('b) => compare('a);

* Alias of by

module Contravariant: BsBastet.Interface.CONTRAVARIANT with type Contravariant.t('a) = compare('a);

include { ... };

let reverse: a. compare('a) => compare('a);

* Creates a new compare function that is the reverse of the given compare function

let compareAsIntBy: a. compare('a) => 'a => 'a => int;

* Compares two values using the given compare function

let compareAsInt: (module BsBastet.Interface.ORD with type t = 'a) => 'a => 'a => int;

* Compares two values using the given ORD module

let minBy: a. compare('a) => 'a => 'a => 'a;

* Finds the minimum of two values using the given compare function

let min: (module BsBastet.Interface.ORD with type t = 'a) => 'a => 'a => 'a;

* Finds the minimum of two values using the given ORD module

let maxBy: a. compare('a) => 'a => 'a => 'a;

* Finds the maximum of two values using the given compare function

let max: (module BsBastet.Interface.ORD with type t = 'a) => 'a => 'a => 'a;

* Finds the maximum of two values using the given ORD module

let lessThanBy: a. compare('a) => 'a => 'a => bool;

* Indicates if the item on the left is less than the item on the right using the given compare function

let ltBy: compare('a) => 'a => 'a => bool;

* Alias for `lessThanBy`

let lessThan: (module BsBastet.Interface.ORD with type t = 'a) => 'a => 'a => bool;

* Indicates if the item on the left is less than the item on the right using the given ORD module

let lt: (module BsBastet.Interface.ORD with type t = 'a) => 'a => 'a => bool;

* Alias for `lessThan`

let lessThanOrEqBy: a. compare('a) => 'a => 'a => bool;

* Indicates if the item on the left is less than or equal to the item on the right using the given compare function

let lteBy: compare('a) => 'a => 'a => bool;

* Alias for lessThanOrEqBy`

let lessThanOrEq: (module BsBastet.Interface.ORD with type t = 'a) => 'a => 'a => bool;

* Indicates if the item on the left is less than or equal to the item on the right using the given ORD module

let lte: (module BsBastet.Interface.ORD with type t = 'a) => 'a => 'a => bool;

* Alias for `lessThanOrEq`

let greaterThanBy: a. compare('a) => 'a => 'a => bool;

* Indicates if the item on the left is greater than the item on the right using the given compare function

let gtBy: compare('a) => 'a => 'a => bool;

* Alias for `greaterThanBy`

let greaterThan: (module BsBastet.Interface.ORD with type t = 'a) => 'a => 'a => bool;

* Indicates if the item on the left is greater than the item on the right using the given ORD module

let gt: (module BsBastet.Interface.ORD with type t = 'a) => 'a => 'a => bool;

* Alias for `greaterThan`

let greaterThanOrEqBy: a. compare('a) => 'a => 'a => bool;

* Indicates if the item on the left is greater than or equal to the item on the right using the given compare function

let gteBy: compare('a) => 'a => 'a => bool;

* Alias for `greaterThanOrEqBy`

let greaterThanOrEq: (module BsBastet.Interface.ORD with type t = 'a) => 'a => 'a => bool;

* Indicates if the item on the left is greater than or equal to the item on the right using the given ORD module

let gte: (module BsBastet.Interface.ORD with type t = 'a) => 'a => 'a => bool;

* Alias for `greaterThanOrEq`

let clampBy: a. compare('a) => min:'a => max:'a => 'a => 'a;

Ensure a provided value falls between a max and min (inclusive). Note that if the provided min is greater than the provided max, the max is always returned. This is considered an incorrect use of `clamp`.

``` let clamp = clampBy(Int.compare); clamp(~min=0, ~max=5, 3) == 3; clamp(~min=0, ~max=5, 0) == 0; clamp(~min=0, ~max=3, 4) == 3; clamp(~min=1, ~max=0, 2) == 0; // don't do this ```

let clamp: (module BsBastet.Interface.ORD with type t = 'a) => min:'a => max:'a => 'a => 'a;

* First-class module version of clampBy

let betweenBy: a. compare('a) => min:'a => max:'a => 'a => bool;

Determine whether a provided value falls between a min and max.

let between: (module BsBastet.Interface.ORD with type t = 'a) => min:'a => max:'a => 'a => bool;

* First-class module version of betweenBy

let abs: (module BsBastet.Interface.ORD with type t = 'a) => (module BsBastet.Interface.RING with type t = 'a) => 'a => 'a;

Absolute value: if x is gte zero, return zero, otherwise negate x.

let signum: (module BsBastet.Interface.ORD with type t = 'a) => (module BsBastet.Interface.RING with type t = 'a) => 'a => 'a;

Sign function, evaluates to one for values >= zero, and negative one for values less than zero.