Module Relude_SequenceZipper

module WithSequence: (S: Relude_Interface.SEQUENCE) => { ... };

* Creates a Zipper using the given SEQUENCE as the backing implementation * * Heavily inspired by Queensland Function Programming Lab Haskell implementation, * although without many of the advanced capabilities, like the ListZipperOp stuff. * https://github.com/qfpl/list-zipper/blob/master/src/Data/ListZipper.hs * * See also this very enlightening presentation about zippers for more background: * http://data.tmorris.net/talks/zippers/0a1062fd0526d7ac1f41ade1e4db1465d311b4fd/zippers.pdf