Module Relude_Timer

let delay: delayMS:int => (unit => unit) => unit => unit;

* Delays the invocation of a function by `delayMS` milliseconds, and returns a function to cancel * the scheduled call.

let repeat: delayMS:int => (unit => unit) => unit => unit;

* Repeats a function every `delayMS` milliseconds, and returns a function to cancel the repeat.

let repeatTimes: delayMS:int => times:int => (unit => unit) => unit => unit;

* Repeats a function every `delayMS` milliseconds, up to `times` times, and returns a function to cancel the * repeat.