Module Relude_Tuple2

let make: a b. 'a => 'b => ('a, 'b);

* Constructs a tuple-2 from 2 values

let fromArray: a. array('a) => option(('a, 'a));

* Constructs a tuple-2 from an array of exactly 2 values

let fromArrayAtLeast: a. array('a) => option(('a, 'a));

* Constructs a tuple-2 from an array of at least 2 values

let fromList: a. list('a) => option(('a, 'a));

* Constructs a tuple-2 from a list of exactly 2 values

let fromListAtLeast: a. list('a) => option(('a, 'a));

* Constructs a tuple-2 from a list of at least 2 values

let showBy: a b. ('a => string) => ('b => string) => ('a, 'b) => string;

* Shows a tuple ('a, 'b) using show functions for 'a and 'b

let eqBy: a b. ('a => 'a => bool) => ('b => 'b => bool) => ('a, 'b) => ('a, 'b) => bool;

* Compares two tuples of type ('a, 'b) for equality using equality functions for 'a and 'b

module WithEqs: (EqA: BsBastet.Interface.EQ) => (EqB: BsBastet.Interface.EQ) => { ... };

* Creates a module containing helpers and an Eq module and extensions for the type (a, b)

module type EQ_BY_F = (EqA: BsBastet.Interface.EQ) => (EqB: BsBastet.Interface.EQ) => (A: Relude_Interface.FUNCTION_1 with type EQ_BY_F.b = (EqA.t, EqB.t)) => BsBastet.Interface.EQ with type EQ_BY_F.t = A.a;

* Creates an Eq module for a type t, given an Arrow from t => (a, b) and EQ modules for a and b

module EqBy: EQ_BY_F;

let compareBy: a b. ('a => 'a => Relude_Ordering.t) => ('b => 'b => Relude_Ordering.t) => ('a, 'b) => ('a, 'b) => Relude_Ordering.t;

* Compares two tuples of type ('a, 'b) for ordering, given comparison functions for 'a and 'b * * The 'a value is checked first, and if the 'as are equal, the 'b compare is checked.

module WithOrds: (OrdA: BsBastet.Interface.ORD) => (OrdB: BsBastet.Interface.ORD) => { ... };

* Creates an Ord module for (a, b), given Ord instances for a and b

module type ORD_BY_F = (OrdA: BsBastet.Interface.ORD) => (OrdB: BsBastet.Interface.ORD) => (A: Relude_Interface.FUNCTION_1 with type ORD_BY_F.b = (OrdA.t, OrdB.t)) => BsBastet.Interface.ORD with type ORD_BY_F.t = A.a;

* Creates an Ord instance for a type t, given an arrow from t => (a, b) and Ord instances for a and b

module OrdBy: ORD_BY_F;