Module Relude_Void

type t =

| Void(t)

;

Void.t is a type for which it is impossible to construct a value.

let absurd: t => 'a;

A function that can be used when you need to provide a function from Void.t => 'a. The function will never actually be called, because it's impossible to construct a value of type Void.t.

let show: t => string;

Show for Void.t is absurd because there is no value of type Void.t that can be shown.