Module Relude_WriterT.WriterLog

WriterLog contains a basic typeclass that has a log type t and a Monoid for t for use as a log. Not all writer typeclass instances require a monoid, but I'm going to normalize on it here for simplicity, and to avoid even more module functor noise. This also contains base implementation for List and Array, which just need to be provided the entry type in order to be used.

module type LOG = { ... };

module List: { ... };

Writer log that is backed by a List of entries of type Entry.t

module Array: { ... };

Writer log that is backed by a Array of entries of type Entry.t