Module Relude_Decimal

type mantissa = int;

`Relude.Decimal` contains a type `t` which represents arbitrary precision numeric values, backed by an `int` mantissa and an `int` exponent. This can be useful for representing currency values without loss in precision or floating point errors.

Because the mantissa and exponent are backed by ints, there is a fairly restrictive range of values that can be represented.

```reason let a = Decimal(12345, -2); // 123.45 let b = Decimal(6789, 3); // 6789000 ```

* The type of the base of our Decimal

type exponent = int;

* The type of the exponent of our Decimal

type t =

| Decimal(mantissa, exponent)

;

Represents an arbitrary precision number, backed by an integer mantissa, and integer exponent.

type rounding =

| Truncate

| RoundUp

| RoundDown

| RoundUpOrDown

;

Represents a preference for rounding a `Decimal` value, when applicable.

let make: int => int => t;

Constructs a `Decimal` from a `mantissa` and `exponent`

let fromInt: int => t;

Makes a `Decimal` from an `int`

let show: t => string;

Attempts to parse a `Decimal` from a `string`

Renders the `Decimal` value to a string.

e.g. ``` Decimal(12345, -2) |> Decimal.toString // "123.45" Decimal(12345, 3) |> Decimal.toString // "12345000" ```

let tenToThePowerOfPositive: int => int;

Rounds a `Decimal` using the given rounding strategy.

Computes the value of `10^exponent`.

This return value of this function is undefined for exponent values < 0

E.g. `tenToThePowerOf(3) == 1000`

let normalize: t => t => (t, t, int);

Normalizes the exponent to the minimal exponent for two given `Decimal` values.

E.g. ``` let a = Decimal(12345, -2); // 123.45 let b = Decimal(12345, 3); // 12345000 let res = Decimal.normalize(a, b); > res == (Decimal(12345, -2), Decimal(1234500000, -2), -2) ```

let add: t => t => t;

Adds two `Decimal` values with no attempt at avoiding overflow.

Note: the arguments are in order of `lhs`, `rhs`

let (+..): t => t => t;

Infix operator for `add`

let subtract: t => t => t;

Subtracts two `Decimal` values with no attempt at avoiding overflow.

Note: the arguments are in order of `lhs`, `rhs`

let (-..): t => t => t;

Infix operator for `subtract`