Module Relude_Js_Promise

let toIO: a. Js.Promise.t('a) => Relude_IO.t('a, Js.Promise.error);

Relude.Js.Promise contains utilities for interoperating with Js.Promise. Many of these functions will help you convert to and from Relude.IO.

Lifts a Js.Promise into a Relude.IO

Note: prefer toIOLazy over this function if possible. A Js.Promise is eagerly executed, so using toIO with an already-constructed and running Js.Promise will not suspend the side effects.

let toIOLazy: a. (unit => Js.Promise.t('a)) => Relude_IO.t('a, Js.Promise.error);

Lifts a lazily-executed Js.Promise into a Relude.IO.

let fromIOWithResult: a e. Relude_IO.t('a, 'e) => Js.Promise.t(Pervasives.result('a, 'e));

Converts a Relude.IO into a Js.Promise.t. This function will cause the IO effects to be run.

The promise that is returned will not reject, it will instead have a result as its resolution.

let fromIO: a e. Relude_IO.t('a, 'e) => Js.Promise.t('a);

Converts a Relude.IO into a Js.Promise.t. This function will cause the IO effects to be run.

The error channel is unsafely coerced into the promise error type, which is probably fine, because the Js.Promise error type is opaque.

let fromIOExn: a. Relude_IO.t('a, exn) => Js.Promise.t('a);

Converts a Relude.IO with an extensible OCaml exn as the error type into a Js.Promise.t. This function will cause the IO effects to be run.

let fromIOJsExn: a. Relude_IO.t('a, Js.Exn.t) => Js.Promise.t('a);

Converts a Relude.IO with a Js.Exn.t as the error type into a Js.Promise.t. This function will cause the IO effects to be run.