Module Relude_Result

type t('a, 'e) = Pervasives.result('a, 'e) =

| Ok('a)

| Error('e)

;

let ok: a e. 'a => t('a, 'e);

`ok()` is a synonym for `pure()`.

let error: a e. 'e => t('a, 'e);

`error(x)` wraps its argument in `Error()`.

error("Not even") == Error("Not even");

let unit: e. t(unit, 'e);

Result.unit is an Ok value that holds ().

Result.unit == Ok();

let getOk: a e. t('a, 'e) => option('a);

`getOk(result)` returns `Some(v)` when `result` is of the form `Ok(v)`; otherwise it returns `None`.

getOk(Ok(1066)) == Some(1066);
getOk(Error("bad value")) == None;

let toOption: a e. t('a, 'e) => option('a);

Alias for `getOk`

let getError: a e. t('a, 'e) => option('e);

`getError(result)` returns `Some(e)` when `result` is of the form `Error(e)`; otherwise it returns `None`.

getError(Ok(1066)) == None;
getError(Error("bad value")) == Some("bad value");

let isOk: a e. t('a, 'e) => bool;

`isOK(result)` returns `true` if `result` is of the form `Ok(val)`, `false` otherwise.

let isError: a e. t('a, 'e) => bool;

`isError(result)` returns `true` if `result` is of the form `Error(err)`, `false` otherwise.

let fold: a e c. ('e => 'c) => ('a => 'c) => t('a, 'e) => 'c;

`fold(errFcn, okFcn, x)` returns `okFcn(v)` when `x` is of the form `Ok(v)`; it returns `errFcn(e)` when `x` is of the form `Error(e)`.

let errToInt = (_) => {-1};
let cube = (x) => {x * x * x};
fold(errToInt, cube, Ok(12)) == 1728;
fold(errToInt, cube, Error("bad")) == -1;

let getOrElse: a e. 'a => t('a, 'e) => 'a;

`getOrElse(default, result)` returns `v` when `result` is of the form `Ok(v)`; otherwise, it returns `default`.

let safeAvg = (total, n): Relude.Result.t(float, string) => {
  if (n > 0) {
    Ok(total /. float_of_int(n));
  } else {
    Error("Cannot calcuate average");
  };
};

getOrElse(0.0, safeAvg(32.0, 4)) == 8.0;
getOrElse(0.0, safeAvg(0.0, 0)) == 0.0;

let getOrElseLazy: a e. (unit => 'a) => t('a, 'e) => 'a;

`getOrElseLazy` returns the `Ok` value inside the `result` or calls the provided function to get a value if the result is `Error`. Unlike `getOrElse`, this only constructs the fallback value if it's needed, which may be useful if the fallback is expensive to construct.

let merge: a. t('a, 'a) => 'a;

`merge(x)` “unwraps” its argument. If `x` is of the form `Ok(v)`, the result is `v`. If `x` is of the form Error(e), the result is `e`.

merge(Ok(2)) == 2;
merge(Error("message")) == "message";

let flip: a e. t('a, 'e) => t('e, 'a);

`flip(x)` flips the values between the `Ok` and `Error` constructors. `Ok(val)`.

flip(Ok(3)) == Error(3);
flip(Error(-1)) == Ok(-1);

let compose: a b c e. t('b => 'c, 'e) => t('a => 'b, 'e) => t('a => 'c, 'e);

let andThen: a b c e. t('a => 'b, 'e) => t('b => 'c, 'e) => t('a => 'c, 'e);

let map: a b e. ('a => 'b) => t('a, 'e) => t('b, 'e);

`map(f, x)` returns `Ok(f(x))` if `x` is of the form `OK(v)`. It returns `Error(e)` if `x` is of the form `Error(e)`.

map(x => sqrt(float_of_int(x)), Ok(4)) == Ok(2.0);
map(x => sqrt(float_of_int(x)), Error("bad")) == Error("bad");

One place you might use `map()` is in validating input to an “ordinary” function. Consider the (highly artificial) example where you have a function that will cube a number, but you only want to do so if the number is even. Here are the functions:

type intResult = Relude.Result.t(int, string);
let cube = (x) => {x * x * x};
let testEven = (n): intResult => {
  (n mod 2 == 0) ? Ok(n) :
    Error(string_of_int(n) ++ " is not even.")
};

We can now make calls like this:

map(cube, testEven(12)) == Ok(1728);
map(cube, testEven(5)) == Error("5 is not even.");

This is something we could have done with a simple `if` statement, but we will see `map()` become useful when we have several things to validate. (See `apply()` and `map2()`, `map3()`, etc.)

let mapOk: ('a => 'b) => t('a, 'c) => t('b, 'c);

`mapOk` is a synonym for `map`

let mapError: a e1 e2. ('e1 => 'e2) => t('a, 'e1) => t('a, 'e2);

`mapError(f, x)` returns `Ok(v)` if `x` is of the form `OK(v)`. It returns `Error(f(e))` if `x` is of the form `Error(e)`.

mapError((x) => {"Err: " ++ x}, Ok(4)) == Ok(4);
mapError((x) => {"Err: " ++ x}, Error("bad")) == Error("Err: bad");

let bimap: a b e1 e2. ('a => 'b) => ('e1 => 'e2) => t('a, 'e1) => t('b, 'e2);

`bimap(f, g, x)` returns `Ok(f(v))` if `x` is of the form `Ok(v)`; it returns `Error(g(e))` if `x` is of the form `Error(e)`.

let cube = x => x * x * x;
let label = x => "Err: " ++ x;

bimap(cube, label, Ok(12)) == Ok(1728);
bimap(cube, label, Error("bad")) == Error("Err: bad");

let tap: a e. ('a => unit) => t('a, 'e) => t('a, 'e);

In `tap(f, x)`, function `f()` returns `unit`. Thus, `f()` is used only for its side effects. If `x` is of the form `Ok(v)`, `tap()` calls `f(v)`. The `tap()` function returns the argument `x`.

tap(x => Js.log(x), Ok(4)) == Ok(4); // prints 4
tap(x => Js.log(x), Error("bad")) == Error("bad"); // prints nothing

let tapOk: a e. ('a => unit) => t('a, 'e) => t('a, 'e);

`tapOk` is a synonym for `tap`.

let tapError: a e. ('e => unit) => t('a, 'e) => t('a, 'e);

In `tapError(f, x)`, function `f()` returns `unit`. Thus, `f()` is used only for its side effects. If `x` is of the form `Error(v)`, `tap()` calls `f(v)`. The `tap()` function returns the argument `x`.

tapError(x => Js.log(x), Ok(4)) == Ok(4); // prints nothing
tapError(x => Js.log(x), Error("bad")) == Error("bad"); // prints "bad"

let apply: a b e. t('a => 'b, 'e) => t('a, 'e) => t('b, 'e);

`apply(fcn, x)` provides a way of creating a chain of validation functions.

If `fcn` is of the form `Ok(f)`, function `f` is applied to `x`. If `x` is `Ok(v)`, the result is `Ok(f(v))`. If `x` is `Error(err)`, the error is passed onwards.

If `fcn` is itself of the form `Error(err)` and `x` is `Ok(v)`, `Error(err)` is passed on.

Finally, if both `fcn` and `x` are `Error(e1)` and `Error(e2)`, the result is `Error(e2)`.

Using `apply()` properly is somewhat complex. See the example in the `__tests__/Relude_Validation_test.re` file for more details. (It uses `VOk` and `VError`, but the logic is identical.)

let map2: a b c x. ('a => 'b => 'c) => t('a, 'x) => t('b, 'x) => t('c, 'x);

`map2(f, x, y)` has as its first argument a function that takes two values. If both of `x` and `y` are of the form `Ok(xv)` and `Ok(yv)`, `map2()` returns `Ok(f(xv, yv))`. Otherwise, it returns the last value of type `Error(e)` that it encounters.

Here is another artificial example that concatenates a string and integer, but only if the string is non-empty and the number is odd:

let combine = (str, n) => {str ++ " " ++ string_of_int(n)};

type strResult = Relude.Result.t(string, string);
let testStr = (s): strResult => (s == "") ? Error("empty string"): Ok(s);
let testOdd = (n): intResult => (n mod 2 == 0) ? Error("not odd") : Ok(n);

map2(combine, testStr("cloud"), testOdd(9)) == Ok("cloud 9");
map2(combine, testStr("cloud"), testOdd(10)) == Error("not odd");
map2(combine, testStr(""), testOdd(9)) == Error("empty string");
map2(combine, testStr(""), testOdd(10)) == Error("not odd");

let map3: a b c d x. ('a => 'b => 'c => 'd) => t('a, 'x) => t('b, 'x) => t('c, 'x) => t('d, 'x);

`map3(f, x, y, z)` has as its first argument a function that takes three values. If all of `x`, `y`, and `z` are of the form `Ok(xv)`, `Ok(yv)`, and `Ok(zv)`, `map3()` returns `Ok(f(xv, yv, zv))`. Otherwise, it returns the last value of type `Error(e)` that it encounters.

The following example builds on the example from `map2()` and does not show all the possible combinations.

let combine = (str, n1, n2) => {str ++ " " ++ string_of_int(n1 * n2)};
let testLimit = (n): intResult => {(n < 100) ? Ok(n) : Error("too big")};

map3(combine, testStr("cloud"), testOdd(3), testLimit(3)) == Ok("cloud 9");
map3(combine, testStr("cloud"), testOdd(2), testLimit(3)) == Error("not odd");
map3(combine, testStr(""), testOdd(3), testLimit(3)) == Error("empty string");
map3(combine, testStr(""), testOdd(10), testLimit(100)) == Error("too big");

let map4: a b c d e x. ('a => 'b => 'c => 'd => 'e) => t('a, 'x) => t('b, 'x) => t('c, 'x) => t('d, 'x) => t('e, 'x);

`map4(f, x, y, z, w)` has as its first argument a function that takes four values. If all of `x`, `y`, `z`, and `w` are of the form `Ok(xv)`, `Ok(yv)`, `Ok(zv)`, and `Ok(wv)`, `map4()` returns `Ok(f(xv, yv, zv, wv))`. Otherwise, it returns the last value of type `Error(e)` that it encounters.

This example uses validation functions defined in the example from `map3()`.

let combine = (s1, n2, n3, n4) => {s1 ++ " " ++ string_of_int(n2 + n3 + n4)};
let testPositive = (n): intResult => {(n > 0) ? Ok(n) : Error("not positive")};
map4(combine, testStr("car"), testOdd(49), testPositive(2), testLimit(3)) == Ok("car 54");
map4(combine, testStr("car"), testOdd(50), testPositive(2), testLimit(200)) == Error("too big");
map4(combine, testStr(""), testOdd(49), testPositive(-5), testLimit(0)) == Error("not positive");
map4(combine, testStr(""), testOdd(48), testPositive(-9), testLimit(200))
  == Error("too big"); // all failures

let map5: a b c d e f x. ('a => 'b => 'c => 'd => 'e => 'f) => t('a, 'x) => t('b, 'x) => t('c, 'x) => t('d, 'x) => t('e, 'x) => t('f, 'x);

`map5(f, x, y, z, w, q)` has as its first argument a function that takes five values. If all of `x`, `y`, `z`, `w`, and `q` are of the form `Ok(xv)`, `Ok(yv)`, `Ok(zv)`, `Ok(wv)`, and `Ok(qv)`, `map5()` returns `Ok(f(xv, yv, zv, wv, qv))`. Otherwise, it returns the last value of type `Error(e)` that it encounters.

The following examples do not show all the possible combinations.

let combine = (s1, n2, n3, n4, n5) => {s1 ++ " " ++ string_of_int(n2 + n3 + n4 + n5)};
let testNegative = (n): intResult => {(n < 0) ? Ok(n) : Error("not negative")};

map5(combine, testStr("square"), testOdd(5), testPositive(2),
  testLimit(3), testNegative(-9)) == Ok("square 1");
map5(combine, testStr("square"), testOdd(2), testPositive(5), testLimit(200),
  testNegative(-3)) == Error("too big");
map5(combine, testStr(""), testOdd(3), testPositive(5), testLimit(7),
  testNegative(-9)) == Error("empty string");
map5(combine, testStr(""), testOdd(2), testPositive(-2), testLimit(200),
  testNegative(42)) == Error("not negative"); // all failures

let pure: a e. 'a => t('a, 'e);

`pure(x)` wraps its argument in `Ok()`.

pure(3) == Ok(3);

let bind: a b e. t('a, 'e) => ('a => t('b, 'e)) => t('b, 'e);

In `bind(r, f)`, `f()` is a function that takes a non-`Result` argument and returns a `Result` value. If `r` is of the form `Ok(val)`, then `bind()` returns `f(val)`. Otherwise, it returns `r`, which will be an `Error(err)`.

Note: `bind()` is the same as `flatMap()`, except with the arguments in reverse order.

let safeSqrt = (x): Relude.Result.t(float, string) => {
  (x >= 0.0) ? Ok(sqrt(x)) : Error("cannot be negative")
};
bind(Ok(4.0), safeSqrt) == Ok(2.0);
bind(Error("invalid float"), safeSqrt) == Error("invalid float");

let flatMap: a b e. ('a => t('b, 'e)) => t('a, 'e) => t('b, 'e);

In `flatMap(f, r)`, `f()` is a function that takes a non-`Result` argument and returns a `Result` value. If `r` is of the form `Ok(val)`, then `flatMap()` returns `f(val)`. Otherwise, it returns `r`, which will be an `Error(err)`.

let safeSqrt = x =>
  (x >= 0.0) ? Ok(sqrt(x)) : Error("cannot be negative");

flatMap(safeSqrt, Ok(4.0)) == Ok(2.0);
flatMap(safeSqrt, Error("invalid float")) == Error("invalid float");

Note: `flatMap()` is the same as `bind()`, except with the arguments in reverse order.

let flatten: a. t(t('a, 'e), 'e) => t('a, 'e);

* Flattens a nested Result one time.

let alt: a e. t('a, 'e) => t('a, 'e) => t('a, 'e);

`alt(r1, r2)` takes two `Result` arguments. If both are `Ok(..)`, the first one is returned. If only one is `Ok(..)`, it is returned. If both are `Error(..)`, the last one is returned.

### Example ```re alt(Ok(2), Ok(3)) == Ok(2); alt(Error("bad"), Ok(3)) == Ok(3); alt(Ok(2), Error("worse")) == Ok(2); alt(Error("bad"), Error("worse")) == Error("worse"); ```

let align: a b e. t('a, 'e) => t('b, 'e) => t(Relude_Ior_Type.t('a, 'b), 'e);

* Converts two results into a result where either side or both sides have succeded, or a result that * fails if both sides failed.

let alignWith: a b c e. (Relude_Ior_Type.t('a, 'b) => 'c) => t('a, 'e) => t('b, 'e) => t('c, 'e);

Similar to map2, but captures successful values from either or both sides

let catchError: a e1 e2. ('e1 => t('a, 'e2)) => t('a, 'e1) => t('a, 'e2);

`catchError(f, r)` returns `f(e)` when `r` is of the form `Error(e)`; otherwise it returns `r` (an `Ok` value) unchanged.

let labelMessage = s => "Attn: " ++ s;

catchError(labelMessage, Ok(2)) == Ok(2);
catchError(labelMessage, Error("not even")) == Error("Attn: not even");

let handleError: a e. ('e => 'a) => t('a, 'e) => t('a, Relude_Void.t);

* `handleError(f, r)` converts errors into successful values, and returns a * Result where the error channel is voided, to indicate that the error has * been handled

let mapHandleError: a e b. ('a => 'b) => ('e => 'b) => t('a, 'e) => t('b, Relude_Void.t);

* Maps the success channel and handles an error on the error channel to end up with an * Result of a new type with a voided error channel

let recover: a e. 'a => t('a, 'e) => t('a, 'e);

`recover(goodValue, result)` returns `result` if it is of the form `Ok(..)`. If `result` is of the form `Error(..)`, `recover` returns `Ok(goodValue)`.

let safeAvg = (total, n): Relude.Result.t(float, string) => {
  if (n > 0) {
    Ok(total /. float_of_int(n));
  } else {
    Error("Cannot calcuate average");
  };
};

recover(0.0, safeAvg(32.0, 4)) == Ok(8.0);
recover(0.0, safeAvg(0.0, 0)) == Ok(0.0);

let fromOption: a e. 'e => option('a) => t('a, 'e);

`fromOption(defaultError, opt)` converts a value of the form `Some(v)` to `Ok(v)`, and converts `None` to `Error(defaultError)`.

fromOption("bad value", Some(3)) == Ok(3);
fromOption("bad value", None) == Error("bad value");

let fromOptionLazy: a e. (unit => 'e) => option('a) => t('a, 'e);

`fromOptionLazy(defaultFcn, opt)` converts a value of the form `Some(v)` to `Ok(v)`, and converts `None` to `Error(defaultFcn())`, which takes no arguments.

This is called a *lazy* function because the default value is not calculated unless it is required.

let defaultErr = () => "bad value";
fromOptionLazy(defaultErr, Some(3)) == Ok(3);
fromOptionLazy(defaultErr, None) == Error("bad value");

let eqBy: a e. ('e => 'e => bool) => ('a => 'a => bool) => t('a, 'e) => t('a, 'e) => bool;

`eqBy(errorEq, okEq, a, b)` compares `a` and `b` for equality as follows:

If both are of the form `Ok(..)` and `Ok(..)`, `eqBy` calls `okEq()` with their values and returns a boolean depending on whether they are equal or not.

If both are of the form `Error(..)`, `eqBy` calls `errorEq()` with their values and returns a boolean depending on whether they are equal or not.

In all other cases, `eqBy()` returns `false`.

let clockEqual = (c1, c2) => c1 mod 12 == c2 mod 12;
let strEqual = (c1, c2) => c1 == c2;

eqBy(strEqual, clockEqual, Ok(14), Ok(2)) == true;
eqBy(strEqual, clockEqual, Ok(14), Ok(3)) == false;
eqBy(strEqual, clockEqual, Error("not an integer"), Error("not an integer")) == true;
eqBy(strEqual, clockEqual, Error("not an integer"), Error("not positive")) == false;
eqBy(strEqual, clockEqual, Ok(14), Error("not positive")) == false;
eqBy(strEqual, clockEqual, Error("not an integer"), Ok(2)) == false;

let tries: a. (unit => 'a) => t('a, exn);

`tries(f)` calls function `f()`, which takes no arguments and returns some value `retVal`. If the call succeeds, the result of `tries()` is `Ok(retVal)`. If it generates an exception, the value is `Error(exception)`.

tries(() => int_of_string("37")) == Ok(37);
tries(() => int_of_string("four")); // returns an exn

let triesAsString: a. (unit => 'a) => t('a, Js.String.t);

`triesAsString(f)` calls function `f()`, which takes no arguments and returns some value `retVal`. If the call succeeds, the result of `tries()` is `Ok(retVal)`. If it generates an exception, the value is `Error(exceptionStr)`, where `exceptionStr` is the string representation of the exception.

triesAsString(() => int_of_string("37")) == Ok(37);
triesAsString(() => int_of_string("four")) ==
  Error("Failure,-2,int_of_string");

let toValidation: Pervasives.result('a, 'b) => Relude_Validation.t('a, 'b);

`toValidation(result)` converts `Ok(val)` to `VOk(val)` and `Error(err)` to `VError(err)`.

let fromValidation: Relude_Validation.t('a, 'b) => Pervasives.result('a, 'b);

`fromValidation(vResult)` converts `VOk(val)` to `Ok(val)` and `VError(err)` to `Error(err)`.

let toValidationNel: t('a, 'e) => Relude_Validation.t('a, Relude_NonEmpty.List.t('e));

`toValidationNel(vResult)` converts `Ok(val)` to `VOk(val)` and `Error(err)` to `VError(err)`, where the list is of type `Relude.NonEmpty.List`.

### Example ```re toValidationNel(Ok(1066)) == Relude.Validation.VOk(1066); toValidationNel(Error("not odd")) == Relude.Validation.VError( Relude.NonEmpty.List.pure("not odd")); ```

You use this function when you have a `Result` type that you wish to use with `Validation` in order to accumulate a list of errors.

let toValidationNea: t('a, 'e) => Relude_Validation.t('a, Relude_NonEmpty.Array.t('e));

`toValidationNea(vResult)` converts `Ok(val)` to `VOk(val)` and `Error(err)` to `VError(|err|)`, where the array is of type `Relude.NonEmpty.Array`.

### Example ```re toValidationNea(Ok(1066)) == Relude.Validation.VOk(1066); toValidationNea(Error("not odd")) == Relude.Validation.VError( Relude.NonEmpty.Array.pure("not odd")); toValidationNea(Error("not odd")); ```

You use this function when you have a `Result` type that you wish to use with `Validation` in order to accumulate an array of errors.

module Bifunctor: BsBastet.Interface.BIFUNCTOR with type Bifunctor.t('a, 'e) = t('a, 'e);

let bimap: ('a => 'b) => ('c => 'd) => Bifunctor.t('a, 'c) => Bifunctor.t('b, 'd);

include { ... };

let mapLeft: ('a => 'c) => Bifunctor.t('a, 'b) => Bifunctor.t('c, 'b);

let mapRight: ('b => 'd) => Bifunctor.t('a, 'b) => Bifunctor.t('a, 'd);

let mapError: ('a => 'b) => Bifunctor.t('c, 'a) => Bifunctor.t('c, 'b);

module Bifoldable: BsBastet.Interface.BIFOLDABLE with type Bifoldable.t('a, 'e) = t('a, 'e);

let bifoldLeft: ('a => 'b => 'a) => ('a => 'c => 'a) => 'a => Bifoldable.t('b, 'c) => 'a;

let bifoldRight: ('a => 'b => 'b) => ('c => 'b => 'b) => 'b => Bifoldable.t('a, 'c) => 'b;

include { ... };

module WithError: (E: BsBastet.Interface.TYPE) => { ... };

* Because Result is a bi-functor, we need to capture the error type in order * to implement many of the single-type-parameter typeclasses. Doing it like * this allows us to unlock a bunch of stuff at once using a single module * functor.